==> 01a_FACS_RSG_gate_checker.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

# input folder
os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_RSG_FACS_checker')

# folder to save output files will create it if it does not exist
outdir="/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_RSG_FACS_checker"
if os.path.isdir(outdir):
    print(outdir, "already exists.")
else:
    os.mkdir(outdir)
    print(outdir, "is created.")

matplotlib.__version__

### Description
Using the three csv files that contain all the events that passed all of the gates I need to create a new column of "Passed_Gates" and then save a csv file with all the SAGs in the summary files folder, and another file of just the ones that passed the gating thresholds. 

SAG_data=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                    index_col=0)

SAG_data

gate181030=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_RSG_FACS_checker/export_CUsersOperatorDesktopFCS Database2018SCGC_2018-10-30AH-545_RS_EBB-Dock_RSG_2018-10-30_001_692 531 Gate.csv')

gate190402=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_RSG_FACS_checker/export_CUsersOperatorDesktopFCS Database2019SCGC_2019-04-02AH-657_EBB-Dock_2019-04-02_RSG_001_692 531 Gate.csv')

gate190709=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_RSG_FACS_checker/export_CUsersOperatorDesktopFCS Database2019SCGC_2019-07-09AH-700_EBB-Dock_2019-07-09_RSG_001_692 531 Gate.csv')
for col in gate181030:
    print(col)

SAG_data['gate']

RSG=SAG_data[SAG_data['gate']=='Respiring cells (RSG-positive)']
dates=[181030, 190404, 190709]
RSG=RSG[RSG['Sample_collection_Date']==RSG['FACS_date']]
RSG=RSG[RSG['Sample_collection_Date'].isin(dates)]
RSG_1810=RSG[RSG['Sample_collection_Date']==181030]
RSG_1810=RSG_1810.rename(columns={'X531.40..488.':'*531_40 [488]', 'X572.27..488.':'*572_27 [488]', 
                             'X692.40..488.':'*692_40 [488]', 'Trigger.Pulse.Width':'Trigger Pulse Width'})

for col in RSG_1810:
    print(col)

passing_RSG_SAGs=RSG_1810.merge(gate181030, how='inner', on=['*531_40 [488]', '*572_27 [488]', '*692_40 [488]', 
                                                             'Trigger Pulse Width', 'FSC', 'SSC'])

passing_RSG_SAGs
