==> 02a_GORG_recruitment_read_counting_per_gen_and_clust_simpledict.ipynb <==
April 9th 2021 Jacob copied this notebook from what Julia had previously done to redo this using the GTDB classification results.

### This notebook takes ~15 hours to run with 32 CPUs and 64G of memory. 
It should be run via a screen session on charlie

## Convoluted description of task

Basically, the 6th column () contains all of the best matches if there are multiple best matches. As long as all of these contigs/SAGs are assigned to the same Genus I can go ahead and classify that read as originating from that Genus. Otherwise, it is of ambiguous origin and I can not assign it to that Genus. I need some assistance in figuring out how to parse the outputs from the DNA and RNA recruitments so that instead of just looking at the Genus (the last column) it looks at the Genus, the contigs that are in the 6th column and confirms that all of the SAGs with contigs are part of that Genus. I have skimmed a couple of the output files and there are defiantly some examples of reads that have best matches to multiple SAGs like the example below from this file /mnt/scgc/simon/microg2p/analyses/GORG_recruitment/no_GEN/GEN_SAGs/results/annotations/All_190709_GoM_RNA_seq_bbmerge_reads_annotated_classified2.txt. 

The genus by data csv file is not sufficient for this. For each date I need a csv file that is a grid of the gene name (prokka_gene) by genus. That way I can differentiate expression between different genes at different time points sorry for this omission I completely spaced on that earlier.

## Data to use
Right now there are 2 sets of recruitment that I need to do this on. The transcriptomic \*annotated.txt files in this folder /mnt/scgc/simon/microg2p/analyses/GORG_recruitment/no_GEN/GEN_SAGs/results/annotations

And the DNA recruitment \*annotated.txt files that you previously created the read_counting notebook to process in this folder /mnt/scgc/simon/microg2p/analyses/GORG_recruitment/201130_GORG_DNA_recruitment/results/annotations

This file should contain all of the SAGs as well as the Genus that they are assigned to /mnt/scgc/simon/microg2p/analyses/ani/All_GoM_SAGs_1cell_20kb_decon_clusters_added.csv

## TODO

* Count reads recruited to each gene name + functional description per Genus, 
* Separate count by whether read recruited exclusively to given Genus, or recruited to several Genera (aka shared with other genera).  
* Count reads recruited per Genus, per functional category, per metagenome.

The plan:
* Dataframe per metagenome will have the columns: ['gene_id','GEN','Exclusive','Shared'], 
* Include ec_number and prokka functions as well
* 'Exclusive' and 'Shared' columns will be read counts


import glob
from collections import defaultdict
import pandas as pd
import os.path as op
import math
import os 
from collections import Counter

import sqlite3
import sqlalchemy

def safe_makedir(dname):
    """
    Make a directory if it doesn't exist, handling concurrent race conditions.
    """
    if not dname:
        return dname
    num_tries = 0
    max_tries = 5
    while not os.path.exists(dname):
        try:
            os.makedirs(dname)
        except OSError:
            if num_tries > max_tries:
                raise
            num_tries += 1
            time.sleep(2)
    return dname

read_files = glob.glob('/mnt/scgc/simon/microg2p/analyses/GORG_recruitment/no_GEN/GEN_SAGs/results/annotations/*annotated.txt') + \
glob.glob('/mnt/scgc/simon/microg2p/analyses/GORG_recruitment/201130_GORG_DNA_recruitment/results/annotations/*annotated.txt')

read_files

Buliding a master dataframe of all GoM ORFs

# file that indicates which sags belong to which category

cat_file = '/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv'
namecolumnid = 'name'
categorycolumnid = 'GTDB_classification'

#catdf_file = '/mnt/scgc/simon/microg2p/analyses/ani/All_GoM_SAGs_1cell_20kb_decon_clusters_added.csv'
catdf = pd.read_csv(cat_file)


gadict = {l[namecolumnid]:l[categorycolumnid] for i, l in catdf.iterrows()}

mmseqs_cluster_file = '/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/mmseqs/analyses/GoM_sag_orfs_80minid_m80.tsv'

head = ['gene_cluster_rep', 'gene']
gcdf = pd.read_csv(mmseqs_cluster_file,
                         sep='\t', names=head)

gcdf['gene_cluster_rep'] = ["_".join(i.split("_")[:-1]) for i in gcdf['gene_cluster_rep']]
gcdf['gene'] = ["_".join(i.split("_")[:-1]) for i in gcdf['gene']]

#gcdict = {}
#gcdict = {l['gene']:l['gene_cluster_rep'] for i, l in gcdf.iterrows()}
### Use the gadict from above to check if genes are part of multiple genera 

gcdf['gene_sag'] = [i.split("_")[0] for i in gcdf['gene']]
gcdf['gene_genus'] = [gadict[i] for i in gcdf['gene_sag']]

# this counts the number of genes represented in each gene cluster
genus_per_cluster = gcdf[gcdf['gene_genus'] != 'Unclassified'].drop_duplicates(subset=['gene_cluster_rep','gene_genus']).groupby('gene_cluster_rep', as_index=False)['gene'].count().rename(columns={'gene':'genus_count'})

genus_per_cluster['gene_cluster_genus_status'] = ['exclusive' if i == 1 else 'shared' for i in genus_per_cluster['genus_count']]
gcdf2 = gcdf.merge(genus_per_cluster[['gene_cluster_rep','gene_cluster_genus_status']], how='left')

gcdf2['gene_cluster_genus_status'] = gcdf2['gene_cluster_genus_status'].fillna('unclassified_only')



Counter(gcdf2['gene_cluster_genus_status']).most_common()

gcdf2.head()

gcdf2['gene_cluster-gene_genus'] = ['{}-{}'.format(l['gene_cluster_rep'],l['gene_genus']) for i, l in gcdf2.iterrows()]
gcdf2.to_csv("/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/mmseqs/analyses/GoM_sag_orfs_80minid_m80_gtdb_added.csv", index=False)

After the above step, you can submit the qsub for the \*simpledict.py script, but before you do that, if you've made changes to the above table, you will want to go into that script and change the intermediate destination directory.  I've done that for the changes made on 7/20/21... the new intermediate output directory is: ```/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210720_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts```


Below is just development scripting that does not need to be run if you run the associated python script.

d2 = {l['gene']:'{}--{}'.format(l['gene_cluster_rep'],l['gene_genus']) for i, l in gcdf2.iterrows()}

We are adding an additional piece of information the output files separated by genera.  We will add the gene cluster(s) associated with each gene hit in a new column.  If so, write that gene cluster id to the intermediate output file. Whether or not the gene cluster is shared or exclusive can be determined by merging the table we created above to the final output.

!head -20 {read_files[0]}

## testing out method...
intermediate_outdir = safe_makedir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3')
#final_outdir = safe_makedir("/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/MMseq_cluster_summaries")

columns=['genus',
        'ec_number', 
        'prokka_function',
        'prokka_gene',
        'seq_ids_lca',
        'exclusive',
        'shared',
        'gene_clusters',
        'gene_clusters_hit_count']

for infile in read_files[:1]:
    
    inid = op.basename(infile).split(".")[0]
    
    #final_gc_outfile = op.join(final_outdir, '{}_reads_by_gen_ko_and_gene_clust.csv'.format(inid))
    
    #if op.exists(final_gc_outfile):
    #    print("output for {infile} already exists, output found at {final_outfile}".format(infile=infile, final_outfile=final_outfile))
    #    continue
    #else:
    reads = 0
    recorded = 0
    with open(infile) as ih:

        outfiles = []
        if op.exists(intermediate_outdir):
            !rm -rf {op.join(intermediate_outdir, inid)}

        outdir = safe_makedir(op.join(intermediate_outdir, inid))

        for j, l in enumerate(ih):

            if j == 0:
                cids = l.strip().split("\t")

            if l.startswith('C'):
                reads += 1
                toks = dict(zip(cids, l.strip().split("\t")))
                nodes_hit = toks['sequence_ids_lca'][:-1].split(",")
                
                labels = [d2[i] for i in nodes_hit]
                gens = [i.split("--")[-1] for i in labels]
                gene_clusters = ",".join(list(set([i.split("--")[0] for i in labels])))

                if len(gens) == 0:
                    keep = l
                    break

                if len(set(gens)) > 1:
                    category = 'shared'
                    shared = 1
                    exclusive=0
                else:
                    category = 'exclusive'
                    exclusive = 1
                    shared = 0

                for gen in list(set(gens)):
                    recorded += 1
                    outfile = op.join(outdir, "{}.tsv".format(gen))
                    if not op.exists(outfile):
                        outfiles.append(outfile)
                        with open(outfile, "w") as oh:
                            print("\t".join(columns), file = oh)
                            print(gen, toks['ec_number'], toks['prokka_function'], toks['prokka_gene'], toks['sequence_ids_lca'], exclusive, shared, gene_clusters, sep="\t", file = oh)
                    else:
                        with open(outfile, "a") as oh:
                            print(gen, toks['ec_number'], toks['prokka_function'], toks['prokka_gene'], toks['sequence_ids_lca'], exclusive, shared, gene_clusters, sep="\t", file = oh)
            if j % 1000000 == 0:
                print(j, 'lines processed')

# to group/count/summarise results, do something like this...

intermediate_outdir = '/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3'
#final_outdir = safe_makedir("/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/MMseq_cluster_summaries")


for infile in read_files:
    inid = op.basename(infile).split(".")[0]
    
    outdir = op.join(intermediate_outdir, inid)
    outfiles = glob.glob(op.join(outdir, '*.tsv'))
    outdf = pd.concat([pd.read_csv(i, sep = "\t") for i in outfiles]).groupby(['genus','gene_clusters'], 
                                                                              as_index=False)['ec_number'].count().rename(columns={'ec_number':'reads_hit'}).sort_values(by='reads_hit', ascending=False)
    outdf['read_library'] = inid
