==> 03_metagenome_derived_GEN_abundance_using_exclusive_reads.ipynb <==
Calculate the number of cells per ml for the November 2017, 2018, and 2019 timepoints using the number of reads recruited to the GENs CDS fraction, estimate the number of reads that would be expected based on the CDS fraction, calculate abundance, normalize abundance by genome size, and recalculate the abundance. Finally multiply by the total number of cells per ml to get the number of cells per ml for each GEN

### Inputs
/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv
/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/summaries/ALL_*_contf_pe_bbmerge_reads_annotated_reads_by_gen_and_ko.csv
/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/GTDB_GoM_SAG_classification_key.csv
### Outputs
/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_Metagenome_analysis/DNA_metagenome_calculated_Genus_abundance.csv
/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_Metagenome_analysis/GoM_metagenome_calculated_Genus_cells_per_ml.csv


import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

os.chdir('/mnt/scgc/simon/microg2p/analyses/GORG_recruitment/210121_summary_tables')
matplotlib.__version__

#### I need to think a liuttle bit about figuring out how to adjust for the number of reads that are shared among multiple genera. Right now I am thinking that the best method is goint to be to:
1. calculate the total number of reads that have been assigned 
2. calculate the number of reclusive reads
3. calculate the proportion of reads that were exclusivelly assigned
4. Use that proportion of reads to adjust down the "total" number of reads per metagenome.

# create table of average genome size per GEN
fields=['name', 'final_assembly_length', 'checkM_estimated_completeness', 'GTDB_classification', 
        'hypothetical_cds_fraction', 'total_CDS_count', 'average_cds_length']
SAG_data=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv', 
                     usecols=fields)
SAG_data['CDS_length']=(SAG_data['total_CDS_count']*SAG_data['average_cds_length'])

SAG_data['GTDB_classification']=SAG_data['GTDB_classification'].fillna('Unclassified')
SAG_data=SAG_data.rename(columns={'GTDB_classification':'genus'})

SAG_data['CDS_fraction']=SAG_data['CDS_length']/SAG_data['final_assembly_length']
SAG_data['checkM_estimated_completeness']=SAG_data['checkM_estimated_completeness'].replace(0, np.nan)
SAG_data=SAG_data[SAG_data['checkM_estimated_completeness'].notna()]
#SAG_data=SAG_data[SAG_data['group72.5'].notna()]
SAG_data['Estimated_genome_size']=SAG_data['final_assembly_length']*(100/SAG_data['checkM_estimated_completeness'])
SAG_data['Estimated_total_CDS_len']=SAG_data['CDS_fraction']*SAG_data['Estimated_genome_size']
SAG_data['Cell_count']=1

SAG_data.head(6)


CDS_frac=pd.pivot_table(SAG_data, values='CDS_fraction', 
                        columns='genus', aggfunc=np.mean)
genome_size=pd.pivot_table(SAG_data, values='Estimated_genome_size', aggfunc=np.mean, columns='genus')


CDS_frac=CDS_frac.append(genome_size)
CDS_frac.loc['CDS_Length_per_cell']=CDS_frac.loc['Estimated_genome_size']*CDS_frac.loc['CDS_fraction']

CDS_frac

#There are 7 GENs that only have SAGs with an estimated completeness of 0. I want to try and double check this
#But all I do is drop rows where estimated completeness is 0 and I lose 7 GENs. 
#GENs 149, 144

CDS_frac_Tran=CDS_frac.T

CDS_frac_Tran=CDS_frac_Tran.rename_axis('genus')
CDS_frac_Tran=CDS_frac_Tran.reset_index()
CDS_frac_Tran['genus']=CDS_frac_Tran['genus'].astype(str)
CDS_frac_Tran=CDS_frac_Tran.sort_values(by=['genus'])
CDS_frac_Tran.dtypes

# The total number of reads recruited per metagneome is:
# this was calculated by counting the number of lines that begun with "C" in the gorg out files
recruit190709=43267551
recruit190402=46925864
recruit181030=45815574
recruit171102=39446211

#
DNA_df171102=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/summaries/ALL_20171102_contf_pe_bbmerge_reads_annotated_read_count_by_genus.csv')
DNA_df171102_pivot=pd.pivot_table(DNA_df171102, index='genus', values='exclusive', aggfunc=np.sum)
DNA_df171102_pivot=DNA_df171102_pivot.rename(columns={'exclusive':'171102_exclusive_reads'})
DNA_df171102_pivot=DNA_df171102_pivot.reset_index()

#calculate the fraction of the total number of reads assigned that are assigned to only a single genus
exclusive_frac1711=DNA_df171102['exclusive'].sum()/recruit171102


abundance_df=CDS_frac_Tran.merge(DNA_df171102_pivot, on='genus', how='outer')
abundance_df

DNA_df181030=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/summaries/ALL_20181030_contf_pe_bbmerge_reads_annotated_read_count_by_genus.csv')
DNA_df181030_pivot=pd.pivot_table(DNA_df181030, index='genus', values='exclusive', aggfunc=np.sum)
DNA_df181030_pivot=DNA_df181030_pivot.rename(columns={'exclusive':'181030_exclusive_reads'})
DNA_df181030_pivot=DNA_df181030_pivot.reset_index()


exclusive_frac1810=DNA_df181030['exclusive'].sum()/recruit181030

abundance_df=abundance_df.merge(DNA_df181030_pivot, on='genus', how='outer')

DNA_df190402=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/summaries/ALL_20190402_contf_pe_bbmerge_reads_annotated_read_count_by_genus.csv')
DNA_df190402_pivot=pd.pivot_table(DNA_df190402, index='genus', values='exclusive', aggfunc=np.sum)
DNA_df190402_pivot=DNA_df190402_pivot.rename(columns={'exclusive':'190402_exclusive_reads'})
DNA_df190402_pivot=DNA_df190402_pivot.reset_index()

exclusive_frac1904=DNA_df190402['exclusive'].sum()/recruit190402


abundance_df=abundance_df.merge(DNA_df190402_pivot, on='genus', how='outer')

DNA_df190709=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/summaries/ALL_20190709_contf_pe_bbmerge_reads_annotated_read_count_by_genus.csv')
DNA_df190709_pivot=pd.pivot_table(DNA_df190709, index='genus', values='exclusive', aggfunc=np.sum)
DNA_df190709_pivot=DNA_df190709_pivot.rename(columns={'exclusive':'190709_exclusive_reads'})
DNA_df190709_pivot=DNA_df190709_pivot.reset_index()

exclusive_frac1907=DNA_df190709['exclusive'].sum()/recruit190709

abundance_df=abundance_df.merge(DNA_df190709_pivot, on='genus', how='outer')


abundance_df=abundance_df.dropna(subset=['171102_exclusive_reads', '181030_exclusive_reads', '190402_exclusive_reads', '190709_exclusive_reads'],axis=0, how='all')
abundance_df.tail(10)

# 
#use the reads per cds to extrapolate the # reads per genome using the CDS_fraction
#figure out the abundance (maybe I am tired but how do i do this when I already did a normalization based on size?)
# normalize the number of reads per genome by genome size


abundance_df['extrapolated_reads_per_genome_20171102']=(abundance_df['171102_exclusive_reads']/abundance_df['CDS_fraction'])*1
abundance_df['extrapolated_reads_per_genome_20181030']=(abundance_df['181030_exclusive_reads']/abundance_df['CDS_fraction'])*1
abundance_df['extrapolated_reads_per_genome_20190402']=(abundance_df['190402_exclusive_reads']/abundance_df['CDS_fraction'])*1
abundance_df['extrapolated_reads_per_genome_20190709']=(abundance_df['190709_exclusive_reads']/abundance_df['CDS_fraction'])*1
abundance_df

# these numbers come from a seqkit stats count of the number of reads in the 
# /mnt/scgc/simon/microg2p/Data/DNA_seq/merged_DNA_seq_runs_201014_201029/*.fastq.gz files

#Adjust the total number of reads so that is proportional to the number of reads that were assigned exclusively to one genus
reads20171102=68439965*exclusive_frac1711
reads20181030=76515787*exclusive_frac1810
reads20190402=73902569*exclusive_frac1904
reads20190709=70871633*exclusive_frac1907



abundance_df.tail(10)

#add the rest of the unrecruited reads to Unclassified
abundance_df.at[295,'extrapolated_reads_per_genome_20171102']=abundance_df.at[295,'extrapolated_reads_per_genome_20171102']+(reads20171102-recruit171102)
abundance_df.at[295,'extrapolated_reads_per_genome_20181030']=abundance_df.at[295,'extrapolated_reads_per_genome_20181030']+(reads20181030-recruit181030)
abundance_df.at[295,'extrapolated_reads_per_genome_20190402']=abundance_df.at[295,'extrapolated_reads_per_genome_20190402']+(reads20190402-recruit190402)
abundance_df.at[295,'extrapolated_reads_per_genome_20190709']=abundance_df.at[295,'extrapolated_reads_per_genome_20190709']+(reads20190709-recruit190709)
abundance_df.tail(10)

abundance_df['total_genome_size_abundance_20171102']=abundance_df['extrapolated_reads_per_genome_20171102']/reads20171102
abundance_df['genome_size_normalized_abundance_20171102']=abundance_df['total_genome_size_abundance_20171102']/abundance_df['Estimated_genome_size']
abundance_df['20171102_genome_size_normalized_metagenome_abundance']=abundance_df['genome_size_normalized_abundance_20171102']/abundance_df['genome_size_normalized_abundance_20171102'].sum()

abundance_df['total_genome_size_abundance_20181030']=abundance_df['extrapolated_reads_per_genome_20181030']/reads20181030
abundance_df['genome_size_normalized_abundance_20181030']=abundance_df['total_genome_size_abundance_20181030']/abundance_df['Estimated_genome_size']
abundance_df['20181030_genome_size_normalized_metagenome_abundance']=abundance_df['genome_size_normalized_abundance_20181030']/abundance_df['genome_size_normalized_abundance_20181030'].sum()

abundance_df['total_genome_size_abundance_20190402']=abundance_df['extrapolated_reads_per_genome_20190402']/reads20190402
abundance_df['genome_size_normalized_abundance_20190402']=abundance_df['total_genome_size_abundance_20190402']/abundance_df['Estimated_genome_size']
abundance_df['20190402_genome_size_normalized_metagenome_abundance']=abundance_df['genome_size_normalized_abundance_20190402']/abundance_df['genome_size_normalized_abundance_20190402'].sum()

abundance_df['total_genome_size_abundance_20190709']=abundance_df['extrapolated_reads_per_genome_20190709']/reads20190709
abundance_df['genome_size_normalized_abundance_20190709']=abundance_df['total_genome_size_abundance_20190709']/abundance_df['Estimated_genome_size']
abundance_df['20190709_genome_size_normalized_metagenome_abundance']=abundance_df['genome_size_normalized_abundance_20190709']/abundance_df['genome_size_normalized_abundance_20190709'].sum()

abundance_df.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_Metagenome_analysis/DNA_metagenome_calculated_Genus_abundance.csv', index=False)
abundance_df

abundance=['genus', '20171102_genome_size_normalized_metagenome_abundance', '20181030_genome_size_normalized_metagenome_abundance', 
           '20190402_genome_size_normalized_metagenome_abundance', '20190709_genome_size_normalized_metagenome_abundance']
df_abundance=abundance_df[abundance].copy()
df_abundance['genus']=df_abundance['genus'].astype(str)
df_abundance

df_abundance['20171102_genome_size_normalized_metagenome_cells/ml']=df_abundance['20171102_genome_size_normalized_metagenome_abundance']*1.55E+06
df_abundance['20181030_genome_size_normalized_metagenome_cells/ml']=df_abundance['20181030_genome_size_normalized_metagenome_abundance']*1.45E+06
df_abundance['20190402_genome_size_normalized_metagenome_cells/ml']=df_abundance['20190402_genome_size_normalized_metagenome_abundance']*1.64E+06
df_abundance['20190709_genome_size_normalized_metagenome_cells/ml']=df_abundance['20190709_genome_size_normalized_metagenome_abundance']*2.92E+06

df_abundance

# I need to redo the key file and run the last two cells after I have done that.

columns=['GTDB_genus', 'full_GTDBtk_classification', '#_of_cells','genus_size_rank']
GoM_GEN_key=pd.read_csv('//mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/GTDB_GoM_SAG_classification_key.csv', usecols=columns)
GoM_GEN_key['GTDB_genus']=GoM_GEN_key['GTDB_genus'].astype(str)

GoM_GEN_key

df_abundance=df_abundance.merge(GoM_GEN_key, left_on='genus', right_on='GTDB_genus', how='left')
df_abundance.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_Metagenome_analysis/GoM_metagenome_calculated_Genus_cells_per_ml.csv', index=False)
df_abundance
