==> 04_AVG_O2_consumption_per_GEN.ipynb <==
rerun on 2021-12-07 after correcting for salinity of media samples

import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob



os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption')
matplotlib.__version__

outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption/'

#!head /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv

pd.set_option("max_rows", 10)


#import RSG SAG data 
columns=['name', 'GTDB_classification', 'Predicted_respiration_rate_fmol_O2perhr', 
         'Sample_collection_Date', 'FACS_date', 'gate']
RSG_SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                    usecols=columns)
RSG_SAGs=RSG_SAGs[RSG_SAGs['Sample_collection_Date']==RSG_SAGs['FACS_date']]
RSG_SAGs['count']=1
RSG_SAGs=RSG_SAGs.rename(columns={'GTDB_classification':'genus', 'Sample_collection_Date':'date'})
RSG_SAGs=RSG_SAGs[RSG_SAGs['gate']=='Respiring cells (RSG-positive)']
RSG_SAGs
avg_RSG=pd.pivot_table(RSG_SAGs, values='Predicted_respiration_rate_fmol_O2perhr', 
                        columns='genus', index='date', aggfunc=np.mean)

avg_RSG=avg_RSG.reset_index()
avg_RSG

#calculate each GEN's percentate at every timepoint
RSG_count_pivot=pd.pivot_table(RSG_SAGs, values='count', 
                        columns='genus', index='date', aggfunc=np.sum)
RSG_count_pivot.sum(axis=1)

RSG_perc=RSG_count_pivot.div(RSG_count_pivot.sum(axis=1), axis=0)
RSG_perc


#replace all genera with <3 fresh RSG labeled SAGS and drop genera that are <3 RSG SAGs at all timepoints
threeRSG_SAGs=RSG_count_pivot[RSG_count_pivot >= 3 ].copy()
threeRSG_SAGs=threeRSG_SAGs.dropna(how='all', axis=1)
threeRSG_SAGs=threeRSG_SAGs.iloc[2:]

threeRSG_SAGs


#OLD CODE THAT USED 1% instead of 3 SAGs

#threeRSG_SAGs_perc=RSG_perc.apply(lambda x: [y if y >=0.01 else 0 for y in x])
#threeRSG_SAGs_perc=threeRSG_SAGs_perc.replace(0, np.nan)
#threeRSG_SAGs_perc=threeRSG_SAGs.div(RSG_count_pivot.sum(axis=1), axis=0)
#threeRSG_SAGs_perc
#threeRSG_SAGs_perc=threeRSG_SAGs_perc.reset_index()
#threeRSG_SAGs_perc=threeRSG_SAGs_perc.iloc[2:]
#threeRSG_SAGs_perc

threeRSG_SAGs_perc=threeRSG_SAGs.div(RSG_count_pivot.sum(axis=1), axis=0)
threeRSG_SAGs_perc=threeRSG_SAGs_perc.iloc[2:]
threeRSG_SAGs_perc=threeRSG_SAGs_perc.reset_index()
threeRSG_SAGs_perc

#Create list of GENs that are >1% for each timepoint
perc_RSG_melt=pd.melt(threeRSG_SAGs_perc, id_vars='date', value_name='relative_abundance')
perc_RSG_melt=perc_RSG_melt.dropna(how='any', axis=0)


df_RSG_GENS_171102=perc_RSG_melt[perc_RSG_melt['date']==171102]
RSG_1perc_GENs_171102=df_RSG_GENS_171102['genus'].values.tolist()

df_RSG_GENS_181030=perc_RSG_melt[perc_RSG_melt['date']==181030]
RSG_1perc_GENs_181030=df_RSG_GENS_181030['genus'].values.tolist()

df_RSG_GENS_190402=perc_RSG_melt[perc_RSG_melt['date']==190402]
RSG_1perc_GENs_190402=df_RSG_GENS_190402['genus'].values.tolist()

df_RSG_GENS_190709=perc_RSG_melt[perc_RSG_melt['date']==190709]
RSG_1perc_GENs_190709=df_RSG_GENS_190709['genus'].values.tolist()

#RSG_1perc_GENs_181030

perc_RSG_melt['date']=perc_RSG_melt['date'].astype(str)


perc_RSG_melt

avg_RSG_melt=pd.melt(avg_RSG, id_vars='date', value_name='Avg_O2_consumed_per_cell')
avg_RSG_melt['date']=avg_RSG_melt['date'].astype(str)
avg_RSG_melt

O2_consumption_3SAG_RSG_Genera=perc_RSG_melt.merge(avg_RSG_melt, how='left', on=['date', 'genus'])
O2_consumption_3SAG_RSG_Genera.to_csv(outdir + 'Average_O2_consumption_for_Genera_with_3+_RSG_SAGs_per_day.csv')

O2_consumption_3SAG_RSG_Genera
