==> 05_KEGG_RNA_transcripts_per_cell.ipynb <==
### This file takes the files that Julia creates as well as the number of mg normalized cells per ml and calculates the number of transcripts per cell for every gene, GEN, date

import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/RNA_summaries/')
outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/RNA_summaries/'

mg_abundance=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_Metagenome_analysis/GoM_metagenome_calculated_Genus_cells_per_ml.csv')
mg_abundance

#Read number to transcript conversion factors 
#These numbers came from a file on my computer and were hard calculated by me
# using the MTST number of reads per MTST standard 
# Users/jmunson-mcgee/Desktop/RNA_misc/MTST_standard_calculations.xlsx 
#Multiply the number of reads mapped to  a transcript in a library this value to get the number of transcripts per ml. 
# I hae already taken into account the sample size of each sample

#Conversion factors
conversion_171102=210.7600209
conversion_181030=67.53617178
conversion_190402=103.6150407
conversion_190709=179.1259172


df_171102=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/summaries/All_171102_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_ko.csv')

df_171102['20171102_transcripts_per_ml']=df_171102['exclusive']*conversion_171102

mg_abundance20171102=mg_abundance[['genus', '20171102_genome_size_normalized_metagenome_cells/ml']]

df_171102=df_171102.merge(mg_abundance20171102, how='left', on='genus')
df_171102['20171102_transcripts_per_cell']=df_171102['20171102_transcripts_per_ml']/df_171102['20171102_genome_size_normalized_metagenome_cells/ml']
df_171102.to_csv(outdir+'All_171102_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_KOnum_transcripts_per_cell.csv')

df_171102

df_181030=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/summaries/All_181030_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_ko.csv')

df_181030['20181030_transcripts_per_ml']=df_181030['exclusive']*conversion_181030

mg_abundance20181030=mg_abundance[['genus', '20181030_genome_size_normalized_metagenome_cells/ml']]

df_181030=df_181030.merge(mg_abundance20181030, how='left', on='genus')
df_181030['20181030_transcripts_per_cell']=df_181030['20181030_transcripts_per_ml']/df_181030['20181030_genome_size_normalized_metagenome_cells/ml']
df_181030.to_csv(outdir+'All_181030_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_KOnum_transcripts_per_cell.csv')

df_181030

df_190402=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/summaries/All_190402_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_ko.csv')

df_190402['20190402_transcripts_per_ml']=df_190402['exclusive']*conversion_190402

mg_abundance20190402=mg_abundance[['genus', '20190402_genome_size_normalized_metagenome_cells/ml']]

df_190402=df_190402.merge(mg_abundance20190402, how='left', on='genus')
df_190402['20190402_transcripts_per_cell']=df_190402['20190402_transcripts_per_ml']/df_190402['20190402_genome_size_normalized_metagenome_cells/ml']
df_190402.to_csv(outdir+'All_190402_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_KOnum_transcripts_per_cell.csv')

df_190402

df_190709=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/summaries/All_190709_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_ko.csv')

df_190709['20190709_transcripts_per_ml']=df_190709['exclusive']*conversion_190709

mg_abundance20190709=mg_abundance[['genus', '20190709_genome_size_normalized_metagenome_cells/ml']]

df_190709=df_190709.merge(mg_abundance20190709, how='left', on='genus')
df_190709['20190709_transcripts_per_cell']=df_190709['20190709_transcripts_per_ml']/df_190709['20190709_genome_size_normalized_metagenome_cells/ml']
df_190709.to_csv(outdir+'All_190709_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_KOnum_transcripts_per_cell.csv')

df_190709


transcripts_20171102=pd.read_csv(outdir+'All_171102_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_KOnum_transcripts_per_cell.csv', 
                                usecols=['genus', 'ko', '20171102_transcripts_per_cell'])
transcripts_20181030=pd.read_csv(outdir+'All_181030_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_KOnum_transcripts_per_cell.csv', 
                                usecols=['genus', 'ko', '20181030_transcripts_per_cell'])
transcripts_20190402=pd.read_csv(outdir+'All_190402_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_KOnum_transcripts_per_cell.csv', 
                                usecols=['genus', 'ko', '20190402_transcripts_per_cell'])
transcripts_20190709=pd.read_csv(outdir+'All_190709_GoM_RNA_seq_bbmerge_reads_annotated_reads_by_gen_and_KOnum_transcripts_per_cell.csv', 
                                usecols=['genus', 'ko', '20190709_transcripts_per_cell'])


transcripts_master=transcripts_20171102.merge(transcripts_20181030, how='outer', 
                                            on=['genus', 'ko', ])
transcripts_master=transcripts_master.merge(transcripts_20190402, how='outer', 
                                            on=['genus', 'ko', ])
transcripts_master=transcripts_master.merge(transcripts_20190709, how='outer', 
                                            on=['genus', 'ko', ])
transcripts_master.to_csv(outdir+'All_GoM_RNA_seq_transcripts_per_cell_by_genus_and_KOnum.csv')

transcripts_master

transcripts_20171102['date']=20171102
transcripts_20171102=transcripts_20171102.rename(columns={'20171102_transcripts_per_cell':'transcripts_per_cell'})

transcripts_20181030['date']=20181030
transcripts_20181030=transcripts_20181030.rename(columns={'20181030_transcripts_per_cell':'transcripts_per_cell'})
transcripts_20190402['date']=20190402
transcripts_20190402=transcripts_20190402.rename(columns={'20190402_transcripts_per_cell':'transcripts_per_cell'})
transcripts_20190709['date']=20190709
transcripts_20190709=transcripts_20190709.rename(columns={'20190709_transcripts_per_cell':'transcripts_per_cell'})

transcripts_20171102

transcripts_stack=transcripts_20171102.append(transcripts_20181030, ignore_index=True)

transcripts_stack=transcripts_stack.append(transcripts_20190402, ignore_index=True)
transcripts_stack=transcripts_stack.append(transcripts_20190709, ignore_index=True)
transcripts_stack=transcripts_stack[transcripts_stack['transcripts_per_cell'].notna()]

transcripts_stack.to_csv(outdir+'All_GoM_RNA_seq_transcripts_per_cell_by_genus_and_KOnum_stacked.csv')
transcripts_stack
