==> 07_gRodon_command_R.ipynb <==
Run gRodon to calculate the growth rates for SAGs
### Citation 
Weissman JL, Hou S, Fuhrman JA. 2020. Estimating maximal microbial growth rates from cultures, metagenomes, and single cells via codon usage patterns. bioRxiv 118:1–10.



library(gRodon)
library(Biostrings)
library(tools)
library(tidyverse)

setwd('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/gRodon')


# output filename
out_name <- "/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/gRodon/gRodon_SAG_growth_rate.csv"

# input ffn directories
ffn_dirs <- c("/mnt/scgc/simon/microg2p/Data/SAG_genes/GoM_190709_genes_ffn/",
              "/mnt/scgc/simon/microg2p/Data/SAG_genes/GoM_190402_genes_ffn/",
              "/mnt/scgc/simon/microg2p/Data/SAG_genes/GoM_181030_genes_ffn/",
              "/mnt/scgc/simon/microg2p/Data/SAG_genes/GoM_171102_genes_ffn/",
              "/mnt/scgc/simon/microg2p/Data/SAG_genes/GoM_170818_genes_ffn/",
              "/mnt/scgc/simon/microg2p/Data/SAG_genes/GoM_170412_genes_ffn/")

#variable that holds all of the outputs
total_pG <- c()

# store warning status and turn off warnings
oldw <- getOption("warn")
options(warn = -1)

#cycle through all the files in the input directories

for (i_dirs in 1:6) {
    ffn_files <- list.files(path=ffn_dirs[i_dirs])
    for (i_files in seq_along(ffn_files)) {
        file_name <- paste(ffn_dirs[i_dirs], ffn_files[i_files], sep="")
        
        # create input for gRodon a list of gene sequences and create TRUE, FALSE list of whether or not gener are "ribosomal proteins"
        genes <- readDNAStringSet(file_name)
        
        # The next couple of lines are new ideas that Jacob is testing
        
        # create a list of the gene name and the length of each gene convert to a list of integers
        lengths <- fasta.seqlengths(file_name)
        lengths <- as.integer(lengths)

        highly_expressed <- grepl("ribosomal protein",names(genes),ignore.case = T)
        
        # use highly_expressed and check if the position of each TRUE value equates to a length >240 in the lengths list.
        # If it does keep it as true, if it does not convert it to FALSE
        highly_expressed <- (highly_expressed) & (lengths > 240)

        
        # if at least 1 gene is "ribosomal protein" run gRodon and store the output I change this to require a certain number of 
        # ribosomal proteins. The article suggests 10 as a good cutoff. I can also filter after the fact using the csv file
        if (sum(highly_expressed) >= 1) {
            pG <- predictGrowth(genes, highly_expressed, mode="partial")
            
            # Store the SAG name, the number of total genes in the ffn file, and the number of "highly expressed ribosomal proteins"
            # in each SAG and add it to the output
            SAG <- file_path_sans_ext(ffn_files[i_files])
            pG[["SAG"]] <- SAG
            pG[["Total_genes"]] <- length(highly_expressed)
            pG[["N_Highly_expressed"]] <- sum(highly_expressed)
            
            # convert output to martix and append it to the total output
            total_pG <- rbind(total_pG, unlist(pG))
        }
    }
}
#turn warnings back on
options(warn = oldw)

#write csv
write.csv(file=out_name, total_pG)


# This code is just here until i get the error fixed and then it will be moved up
write.csv(file=out_name, total_pG)

total_pG
