==> 09a_linear_regression_py_cleanup.ipynb <==
rerun on 2021-12-07 after correcting for salinity of media samples

import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption/Linear_regression')
outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption/Linear_regression/'
matplotlib.__version__


O2_corr=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption/Linear_regression/Gene TCP O2 significant correlations ver 2.csv')
O2_corr=O2_corr.drop(columns='gene')
O2_corr             

colnames=['metabolism', 'path', 'function', 'gene']
Kegg=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Kegg_data/KO_Orthology_ko00001.txt', 
                sep='\t', names=colnames)

Kegg[['ko', 'gene']]=Kegg['gene'].str.split(" ", 1, expand=True)
Kegg
#f[['code','name_of_code']] = df["code"].str.split(" ", 1, expand=True)

merge=O2_corr.merge(Kegg, how='left', on='ko')
merge.to_csv(outdir+ 'Gene TCP O2 significant correlations ver 2_KO_functions.csv')
merge
