==> 10_PCA_RDA_class_seperator.ipynb <==
rerun on 2021-12-07 after correcting for salinity of media samples

import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob

os.chdir("/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption/")
outdir=('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption/')

# read the input file for O2 consumption and split it out by "Class". I might have to merge it with the genus statistiscs file or 
# open that file and use it to create a list of the genera for each class and do a couple of for i in list loops



in_df=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/RNA_summaries/Prokka_Transcripts_percell_O2_percell_10SAG_genera_stats_biogeochemical_master_no2017.csv',
                 index_col=0)

in_df

with open("/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Genera_with_10+_SAGs.txt") as file:
    tenGENs_list = [line.strip() for line in file]

ten_SAG_master=in_df[in_df['genus'].isin(tenGENs_list)]

ten_SAG_master

in_df=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/RNA_summaries/Prokka_znormalized_Transcripts_percell_O2_percell_10SAG_genera_stats_biogeochemical_master_no2017.csv',
                 index_col=0)
norm_ten_SAG_master=in_df[in_df['genus'].isin(tenGENs_list)]

norm_ten_SAG_master

Genus_tax=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Genus_statistics.csv',
                     index_col=0)

classes=Genus_tax.Class.unique()

Genus_tax


MYDIR = ("class_transcripts")
CHECK_FOLDER = os.path.isdir(MYDIR)

# If folder doesn't exist, then create it.
if not CHECK_FOLDER:
    os.makedirs(MYDIR)
    print("created folder : ", MYDIR)

else:
    print(MYDIR, "folder already exists.")
outdir=('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption/class_transcripts/')

ten_SAG_master.dropna(how='all', axis=1, inplace=True)
ten_SAG_master.to_csv(outdir+'All_Genera_with_10+SAGs_Prokka_Transcripts_percell_O2_percell_Genus_stats_biogeochemical_master_no2017.csv')

for i in classes:
    i_df=pd.DataFrame()
    i_df=ten_SAG_master[ten_SAG_master['Class'] == i]
    i=str(i)
    i_df.dropna(how='all', axis=1, inplace=True)
    rows=i_df.shape[0]
    if rows > 0:
        i_df.to_csv(outdir+i+'_Genera_with_10+SAGs_Prokka_Transcripts_percell_O2_percell_Genus_stats_biogeochemical_master_no2017.csv')
    
   # print(i)
i_df

norm_ten_SAG_master.dropna(how='all', axis=1, inplace=True)
norm_ten_SAG_master.to_csv(outdir+'All_Genera_with_10+SAGs_Prokka_znormlized_Transcripts_percell_O2_percell_Genus_stats_biogeochemical_master_no2017.csv')

for i in classes:
    i_df=pd.DataFrame()
    i_df=norm_ten_SAG_master[norm_ten_SAG_master['Class'] == i]
    i=str(i)
    i_df.dropna(how='all', axis=1, inplace=True)
    rows=i_df.shape[0]
    if rows > 0:
        i_df.to_csv(outdir+i+'_Genera_with_10+SAGs_Prokka_znormalized_Transcripts_percell_O2_percell_Genus_stats_biogeochemical_master_no2017.csv')
    
   # print(i)
i_df
