==> 13a_GoM_Syto9_AAI_analysis.ipynb <==
### Using the results from the AAI analysis create graphs of AAI versus O2 consumption for SAG pairs on individual days.

### Inputs
/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv
/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/aai/GoM_RSG_aai_outputs/aai/aai_summary.tsv

### Outputs 
/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/AAI_vs_log10_O2_ratio*.png graph files

rerun on 2021-12-07 after correcting for salinity of media samples The last bit of this code is not finished

import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/aai/')
outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/aai/'
images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/AAI/'
matplotlib.__version__

!head /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv

columns=['name', 'GTDB_classification', 'estimated_diameter', 'Predicted_respiration_rate_fmol_O2perhr',
        'Sample_collection_Date', 'gate', 'FACS_date', 'checkM_estimated_completeness']
SAG_key=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                   usecols=columns)
SAG_key=SAG_key[SAG_key['gate']=='All prokaryotes (SYTO-9-positive)']
SAG_key=SAG_key[SAG_key['Sample_collection_Date']==SAG_key['FACS_date']]
SAG_key=SAG_key[SAG_key['checkM_estimated_completeness']>=50]
SAG_list=SAG_key['name'].tolist()
SAG_key

#read result file from RSG_ANI analysis and parse it to remove AAI=0 and add SAG identifiers

aai=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/aai/GoM_Syto9_outputs/aai/aai_summary.tsv',
                sep='\t')
aai=aai[aai['Mean AAI'] != 0.0]
aai['SAG_A']=aai['Genome A'].str[:10]
aai['SAG_B']=aai['Genome B'].str[:10]

aai=aai[aai['SAG_A'].isin(SAG_list)]
aai=aai[aai['SAG_B'].isin(SAG_list)]


aai

# This bit of code was not finished.

Syto9hist=sns.histplot(data=aai, x='Mean AAI', ax=ax, binwidth=1)
AAIhist=AAIhist.set_title("Sampling Date "+i)
fig=AAIhist.get_figure()
fig.savefig(images+'SYTO9_AAI_histogram_'+".pdf")
