==> 16_Genes_for_Kai.ipynb <==
### inputs
*.txt
/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Transcripts_percell_O2_percell_Genus_stats_biogeochemical_master_no2017.csv
### outputs
*KO_transcripts_per_cell_10_SAG_GENs.csv

rerun on 2021-12-07 after correcting for salinity of media samples

import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/For_kai')
outdir=('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/For_kai/')
matplotlib.__version__

KO_transcripts=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Transcripts_percell_O2_percell_Genus_stats_biogeochemical_master_no2017.csv',
                          index_col=[1,2])
#KO_transcripts=KO_transcripts.drop("Unnamed:0")
KO_transcripts

tax=['full_GTDBtk_classification']

classification=KO_transcripts.filter(tax)
classification

with open('aminopeptidases.txt') as f:
    aminopeptidases = f.read().splitlines()
with open('beta-glucosidases.txt') as f:
    glucosidases = f.read().splitlines()
with open('phosphatases.txt') as f:
    phosphatases = f.read().splitlines()

#aminopeptidases=aminopeptidases.strip('')

glucosidases



#df[df['A'].str.contains("hello")]
#my_list = df.columns.values.tolist()
KO_names=KO_transcripts.columns.values.tolist()
KO_names

amino_names=[i for i in KO_names if i.startswith(tuple(aminopeptidases))]
amino_names
aminopeptidase_KO=KO_transcripts.filter(amino_names)

aminopeptidase_KO=aminopeptidase_KO.dropna(how='all')

aminopeptidase_KO=aminopeptidase_KO.merge(classification, on=['genus', 'date'], how='left')

aminopeptidase_KO.to_csv(outdir+'aminopeptidase_KOs_transcripts_per_cell_10_SAG_Genera.csv')
aminopeptidase_KO


gluco_names=[i for i in KO_names if i.startswith(tuple(glucosidases))]

glucosidases_KO=KO_transcripts.filter(gluco_names)
#glucosidases_KO=glucosidases_KO.dropna(how='all')
glucosidases_KO=glucosidases_KO.merge(classification, on=['genus', 'date'], how='left')

glucosidases_KO.to_csv(outdir+'beta_glucoside_KOs_transcripts_per_cell_10_SAG_Genera.csv')
glucosidases_KO


phospho_names=[i for i in KO_names if i.startswith(tuple(phosphatases))]

phospho_KO=KO_transcripts.filter(phospho_names)
phospho_KO=phospho_KO.dropna(how='all')
phospho_KO=phospho_KO.merge(classification, on=['genus', 'date'], how='left')


phospho_KO.to_csv(outdir+'phosphatase_KOs_transcripts_per_cell_10_SAG_Genera.csv')
phospho_KO
