==> 17_iREP_setup.ipynb <==
URL: https://github.com/christophertbrown/iRep

### Instructions
#### Inputs
iRep requires high-quality draft genomes (>=75% complete, <=175 fragments/Mbp sequence, and <=2% contamination). iRep will only make use of scaffolds ≥5 Kbp in length, thus it is important to make sure that estimates of genome completion are calculated based on scaffolds that pass this threshold

I also want to select just the most complete SAG per genus

The second set of inputs are mapping files in SAM format with the --reorder flag, which should be generated by mapping DNA sequencing reads against assembled genomes using Bowtie2. In this case, provide the path to each SAM file generated for each sample

rerun on 2021-12-07 after correcting for salinity of media samples

import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/iREP')
outdir=('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/iREP')
matplotlib.__version__

SAG_data=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                    index_col=0)
SAG_data=SAG_data[SAG_data['checkM_estimated_completeness'] >= 75]

SAGs75perc=SAG_data['name'].to_list()

MyFile=open('75perc_SAGs.txt','w')

for element in SAGs75perc:
     MyFile.write(element)
     MyFile.write('\n')
MyFile.close()
SAG_data

#If file does not exist do the things to create it
#
if os.path.isfile('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/iREP/75perc_complete_SAGs/75perc_stats.tab')== False:        
    os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/iREP/')
    !grep -f 75perc_SAGs.txt /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/SAG_contigs.txt > 75perc_complete_SAGs/75perc_SAG_contigs.txt

    os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/iREP/75perc_complete_SAGs')

    !while read file; do cp "$file" . ; done < 75perc_SAG_contigs.txt
    !seqkit stats *.fasta -T > 75perc_stats.tab
    os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/iREP/')


stats=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/iREP/75perc_complete_SAGs/75perc_stats.tab',
                 sep='\t')
stats['SAG']=stats['file'].str[:10]
contigs=stats[['num_seqs', 'SAG']].copy()
contigs

SAG_data=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                    index_col=0)
SAG_data=SAG_data[SAG_data['checkM_estimated_completeness'] >= 75]
SAG_data=SAG_data.merge(contigs, how='left', on='SAG', )

SAG_data['frag_per_Mbp']=SAG_data['num_seqs']/(SAG_data['final_assembly_length']/1000000)

SAG_data=SAG_data[SAG_data['frag_per_Mbp'] <= 175]
SAG_data

# I now need to figure out how to get a list of the most complete genome per genus. only one per value
SAG_data.sort_values(by=['GTDB_classification','checkM_estimated_completeness'], ascending = False).drop_duplicates(subset = ['GTDB_classification'], keep='first')

sag_list = list(SAG_data.sort_values(by=['GTDB_classification','checkM_estimated_completeness'], ascending = False).drop_duplicates(subset = ['GTDB_classification'], keep='first')['name'])

genus_SAGs=SAG_data[SAG_data['name'].isin(sag_list)]

genus_SAGs

MyFile=open('largest_SAGs_per_genus.txt','w')

for element in sag_list:
     MyFile.write(element)
     MyFile.write('\n')
MyFile.close()


#copy the files for iREP into a new folder

#!mkdir iREP_SAGs
!grep -f largest_SAGs_per_genus.txt /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/SAG_contigs.txt > largest_SAGs_per_genus_path.txt

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/iREP/iREP_SAGs/')
!while read file; do cp "$file" . ; done < ../largest_SAGs_per_genus_path.txt

!wc -l /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/iREP/largest_SAGs_per_genus_path.txt

I think that at this point I am now at the point where I need to recruit the metagenomes to the genomes for the 78 genera that I have a SAG that fulfills the basic iREP requirements.

### Questions
1. Should I expand this to the largest genome for every genus, recruit to all SAGs? (I think that this would reduce the recruitment of the rest of the SAGs). Recruit to only the ~78 SAGs that I just identified This might result in overrecruitment of reads to these genomes. Does this matter for this analysis?


#!mkdir iRep_SAGs_5kb
!for file in iREP_SAGs/*.fasta; do seqkit seq -m 5000 $file -o ${file/%.fasta/_5kb.fasta}; done
!cp iREP_SAGs/*5kb.fasta iRep_SAGs_5kb/
