==> 19a_hypothetical_protein_clustering.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

def safe_makedir(dname):
    """
    Make a directory if it doesn't exist, handling concurrent race conditions.
    """
    if not dname:
        return dname
    num_tries = 0
    max_tries = 5
    while not os.path.exists(dname):
        try:
            os.makedirs(dname)
        except OSError:
            if num_tries > max_tries:
                raise
            num_tries += 1
            time.sleep(2)
    return dname


os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3/')
outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3/'
images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/hypothetical_genes/'
#tmpdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/tmp/'
matplotlib.__version__

rerun on 2021-12-07 after correcting for salinity of media samples

# Added by Jacob on April 29th 2021 To summarize genes by genus and prokka annotations.
# I added this because after looking over the KO database I realized that proteorhodopsins are very poorly 
# characterized in the database and that if I am interested in those genes I should probably use a different 
# annotation methodology for these

# counting exclusive and shared reads by genus... similar to loop at the end of above cell, except grouping by genus and prokka annotations instead of by genus and ko

final_outdir = safe_makedir("/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3/summaries")
outfiles = glob.glob('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3/All*/') + glob.glob('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3/ALL*/')

for o in outfiles:
    dflist = []
    csvs = glob.glob(op.join(o, '*.tsv'))
    for c in csvs:
        sdf = pd.read_csv(c, sep="\t", engine='python')
        
        sgdf = sdf.groupby(['genus','gene_clusters'])['exclusive','shared'].sum()
        dflist.append(sgdf)

    fdf = pd.concat(dflist).reset_index()
    fdf.to_csv(op.join(final_outdir, '{}_read_count_by_genus_and_prokka_annotations.csv'.format(o.split("/")[-2])), index=False)

vib=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3/ALL_20190709_contf_pe_bbmerge_reads_annotated/Vibrio.tsv',
               sep='\t')

vib.tail()

infiles=glob.glob('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3/summaries/*RNA*.csv')

master=pd.DataFrame()
for i in infiles:
    df=pd.read_csv(i)
    date=i[134:140]


i='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment_mmseq_clusts3/summaries/All_171102_GoM_RNA_seq_bbmerge_reads_annotated_read_count_by_genus_and_prokka_annotations.csv'

test=pd.read_csv(i)
test
date=i[134:140]
pivot=test.pivot_table(index='genus', values='exclusive', columns='gene_clusters', aggfunc=np.sum)
pivot

ex=test['exclusive'].sum()
ex

shared=test['shared'].sum()
shared

perc_excl=(ex/(ex+shared))*100
perc_excl

# List of every that was identified (gene) the gene cluster it belongs to (gene_cluster_rep), the genus, and if it is found in multiple genera or not
annotations=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/mmseqs/analyses/GoM_sag_orfs_80minid_m80_genera_added.csv')
annotations
