==> 20_16S_extraction_and_alignment.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/16S_analysis')
outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/16S_analysis/'
images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/16S_analysis/'
tmpdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/tmp/'

SAG_data=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                     index_col=0)


# generate list of every unique genera
genera=SAG_data['GTDB_classification'].to_list()
genera=list(set(genera))

# generate a txt file of the SAGs that belong to each genus
for i in genera:
    df=pd.DataFrame(columns=SAG_data.columns)
    cond=SAG_data.GTDB_classification == i
    rows=SAG_data.loc[cond, :]
    df=df.append(rows, ignore_index=True)
    SAGs=df.name.to_list()
    MyFile=open('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/lists/'+i+'SAGs.txt', 'w')
    for element in SAGs:
         MyFile.write(element)
         MyFile.write('\n')
    MyFile.close()


# generate a fasta file of evey SSU sequence for each genus

fastafiles=glob.glob('/mnt/scgc/simon/microg2p/Data/SAGs_WGS/SSUs/GoM_SAG_SSUs/*.fasta')
generafiles=glob.glob('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/lists/*SAGs.txt')

# for each genus open the genus txt file, extract the genus name, create a list of SSU files
for file in generafiles:
    fh=open(file, 'r')
    genus=os.path.basename(file)
    genus=genus[:-4]
    fasta=[]
    
    # extract each SAG name
    for line in fh:
        SAG=line.strip('\n')
        
        # check if the each SAG has a SSU fastafile and add that filename to the fasta file list
        for ffile in fastafiles:        
            if SAG in ffile:
                fasta.append(ffile)

    # write fasta file of every SSU assigned to each genus
    with open ('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/16S_analysis/genera_SSU_fasta_files/' + genus + 'SSU.fasta', 'w') as outfile: 
        for item in fasta:
            with open(item) as infile:
                for line in infile:
                    outfile.write(line)

# remove files with a size of 0                    
SSUfasta_files=glob.glob('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/16S_analysis/genera_SSU_fasta_files/*.fasta')
for file in SSUfasta_files:
    if os.path.getsize(file) < 1:
        os.remove(file)

### Question
    1. Create a tree with all ~2000 SSU sequences and collapse nodes
    2. Create a consensus sequence for each genus and create a tree with the ~200 genera and figure out how to indicate the number of SAGs per genus later
