==> 23_per_day_abundnaces.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob
from csv import reader

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files')
outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/'
images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/'
matplotlib.__version__

rerun on 2021-12-07 after correcting for salinity of media samples

syto=pd.read_csv('Syto9_SAG_abundances.csv')

syto.rename(columns={170412:'170412_Syto', 170818:'170818_Syto', 4:'171102_Syto'}, inplace=True)
syto1704=syto[['GTDB_classification', '170412']].copy()
syto1704.columns=['genus', '170412_Syto']

syto1708=syto[['GTDB_classification', '170818']].copy()
syto1708.columns=['genus', '170818_Syto']

syto1711=syto[['GTDB_classification', '171102']].copy()
syto1711.columns=['genus', '171102_Syto']

syto1810=syto[['GTDB_classification', '181030']].copy()
syto1810.columns=['genus', '181030_Syto']

syto1904=syto[['GTDB_classification', '190402']].copy()
syto1904.columns=['genus', '190402_Syto']

syto1907=syto[['GTDB_classification', '190709']].copy()
syto1907.columns=['genus', '190709_Syto']


syto1907


RSG=pd.read_csv('RSG_SAG_abundances.csv')

RSG1704=RSG[['GTDB_classification', '170412']].copy()
RSG1704.columns=['genus', '170412_RSG']

RSG1708=RSG[['GTDB_classification', '170818']].copy()
RSG1708.columns=['genus', '170818_RSG']

RSG1711=RSG[['GTDB_classification', '171102']].copy()
RSG1711.columns=['genus', '171102_RSG']

RSG1810=RSG[['GTDB_classification', '181030']].copy()
RSG1810.columns=['genus', '181030_RSG']

RSG1904=RSG[['GTDB_classification', '190402']].copy()
RSG1904.columns=['genus', '190402_RSG']

RSG1907=RSG[['GTDB_classification', '190709']].copy()
RSG1907.columns=['genus', '190709_RSG']
RSG


lam=pd.read_csv('Laminarin_SAG_abundances.csv')
lam1810=lam[['GTDB_classification', '181030']].copy()
lam1810.columns=['genus', '181030_lam']

lam1904=lam[['GTDB_classification', '190402']].copy()
lam1904.columns=['genus', '190402_lam']

lam1907=lam[['GTDB_classification', '190709']].copy()
lam1907.columns=['genus', '190709_lam']
lam

xyl=pd.read_csv('Xylan_SAG_abundances.csv')

xyl1810=xyl[['GTDB_classification', '181030']].copy()
xyl1810.columns=['genus', '181030_xyl']

xyl1907=xyl[['GTDB_classification', '190709']].copy()
xyl1907.columns=['genus', '190709_xyl']
xyl1907

df1704=syto1704.merge(RSG1704, on='genus', how='outer')
df1704=df1704.dropna(how='all')
df1704.to_csv('170412_genus_abundance.csv')
df1704

df1708=syto1708.merge(RSG1708, on='genus', how='outer')
df1708=df1708.dropna(how='all')
df1708.to_csv('170818_genus_abundance.csv')
df1708

df1711=syto1711.merge(RSG1711, on='genus', how='outer')
df1711=df1711.dropna(how='all')
df1711.to_csv('171102_genus_abundance.csv')

df1711

df1810=syto1810.merge(RSG1810, on='genus', how='outer')
df1810=df1810.merge(lam1810, on='genus', how='outer')
df1810=df1810.merge(xyl1810, on='genus', how='outer')
df1810=df1810.dropna(how='all')
df1810.to_csv('181030_genus_abundance.csv')

df1810

df1904=syto1904.merge(RSG1904, on='genus', how='outer')
df1904=df1904.merge(lam1904, on='genus', how='outer')
df1904=df1904.dropna(how='all')
df1904.to_csv('190402_genus_abundance.csv')

df1904

df1907=syto1907.merge(RSG1907, on='genus', how='outer')
df1907=df1907.merge(lam1907, on='genus', how='outer')
df1907=df1907.merge(xyl1907, on='genus', how='outer')
df1907=df1907.dropna(how='all')
df1907.to_csv('190709_genus_abundance.csv')

df1907

from functools import reduce


data_frames = [df1704, df1708, df1711, df1810, df1904,df1907]
master_abundance=reduce(lambda  left,right: pd.merge(left,right,on=['genus'], how='outer'), data_frames)

master_abundance.to_csv('All_stains_fresh_genus_counts.csv')

master_abundance=master_abundance.fillna(0)
master_abundance

fold=master_abundance[['genus']].copy()

fold['170412_fold_change']=master_abundance['170412_RSG']/master_abundance['170412_Syto']
fold['170818_fold_change']=master_abundance['170818_RSG']/master_abundance['170818_Syto']
fold['171102_fold_change']=master_abundance['171102_RSG']/master_abundance['171102_Syto']
fold['181030_fold_change']=master_abundance['181030_RSG']/master_abundance['181030_Syto']
fold['190402_fold_change']=master_abundance['190402_RSG']/master_abundance['190402_Syto']
fold['190709_fold_change']=master_abundance['190709_RSG']/master_abundance['190709_Syto']

#Notes

#Possible options
#present in both >> color scale (This will be a number)
#present in RSG but not Syto9 >> green (this is represented as "inf") change to appear
#present in Syto9 but not RSG >>red (this is 0.0) change to disappera
#present in neither blank. >> this should be white this is nan

fold=fold.set_index('genus')
fold.replace(to_replace=0, value='Disappear', inplace=True)
fold.replace(to_replace=np.inf, value='Appear', inplace=True)
fold.dropna(how='all', inplace=True, axis=0)
fold.to_csv(outdir+ 'fold_change_Syto9_to_RSG_stain.csv')
