==> 24_genus_characteristics.ipynb <==
The original genus statics file was generated in notebook 08_O2_transcript_biogeochemistry_GEN_stats_master Now I am taking that file and the GEN_statics file and adding a couple of additional values to them.

import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob
from Bio import SeqIO


os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/')
images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/'
outdir=('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/')
matplotlib.__version__

rerun on 2021-12-07 after correcting for salinity of media samples

Genera=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/GEN_statistics.csv', index_col=0)
Genera[['domain', 'phyla', 'class', 'order', 'family', 'genus', 'species']]=Genera['full_GTDBtk_classification'].str.split(';', expand=True)
cols=['GTDB_genus', 'full_GTDBtk_classification', 'domain', 'phyla', 'class', 'order', 'family', 'genus', 'species', 
      '#_of_cells', 'genus_size_rank', 'Average_gc_content', 'Average_hypothetical_cds_fraction', 
      'median_estimated_diameter', 'mean_estimated_genome_length']

Genera=Genera[cols]

Genera[['domain', 'phyla', 'class', 'order', 'family', 'genus', 'species']]=Genera['full_GTDBtk_classification'].str.split(';', expand=True)

cols=['Average_doubling_time', 'median_doubling_time', 'GTDB_genus']
growth=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Genus_statistics.csv', index_col=0,
                  usecols=cols)
growth=growth.drop_duplicates()
Genera=Genera.merge(growth, on='GTDB_genus', how='left')
unc=Genera.iloc[0].copy()
Genera=Genera[Genera['GTDB_genus']!= 'Unclassified']
Genera['domain']=Genera['domain'].str[3:]
Genera['phyla']=Genera['phyla'].str[3:]
Genera['class']=Genera['class'].str[3:]
Genera['order']=Genera['order'].str[3:]
Genera['family']=Genera['family'].str[3:]
Genera['genus']=Genera['genus'].str[3:]
Genera['species']=Genera['species'].str[3:]

Genera=Genera.append(unc)
Genera.to_csv(outdir+'Supplemental_table_genus_statistics.csv')

### Future Additions
1. Add the detection of rhodopsin
2. Add the detection of bacteriochlorophyl


! grep "rhodopsin" /mnt/scgc/simon/microg2p/Data/SAG_genes/faa/GORG_Maine2020_noGEN.fixed.tsv > /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/tmp/rhodopsin.tsv
cols=pd.read_csv('/mnt/scgc/simon/microg2p/Data/SAG_genes/faa/GORG_Maine2020_noGEN.fixed.tsv', sep='\t',  nrows=0).columns.tolist()
rhodopsin=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/tmp/rhodopsin.tsv', sep='\t', names=cols)

cols=['name', 'GTDB_classification']
classification=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv', 
                           usecols=cols)

rhodopsin=rhodopsin.merge(classification, left_on='sag', right_on='name', how='left')

# create a list of all genera where rhodopsin was detected
rhodopsin_genera=rhodopsin[['GTDB_classification']].copy()
rhodopsin_genera=rhodopsin_genera.drop_duplicates()
rhodopsin_genera['Rhodopsin detected']='Yes'
rhodopsin_genera=rhodopsin_genera.rename(columns={'GTDB_classification':'GTDB_genus'})
rhodopsin_genera

! grep "chlorophyll" /mnt/scgc/simon/microg2p/Data/SAG_genes/faa/GORG_Maine2020_noGEN.fixed.tsv > /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/tmp/bacteriochlorophyll.tsv
cols=pd.read_csv('/mnt/scgc/simon/microg2p/Data/SAG_genes/faa/GORG_Maine2020_noGEN.fixed.tsv', sep='\t',  nrows=0).columns.tolist()
bacteriochlorophyll=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/tmp/bacteriochlorophyll.tsv', sep='\t', names=cols)

cols=['name', 'GTDB_classification']
classification=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv', 
                           usecols=cols)

bacteriochlorophyll=bacteriochlorophyll.merge(classification, left_on='sag', right_on='name', how='left')

# create a list of all genera where chlorophyll was detected
bacteriochlorophyll_genera=bacteriochlorophyll[['GTDB_classification']].copy()
bacteriochlorophyll_genera=bacteriochlorophyll_genera.drop_duplicates()
bacteriochlorophyll_genera['Chlorophyll detected']='Yes'
bacteriochlorophyll_genera=bacteriochlorophyll_genera.rename(columns={'GTDB_classification':'GTDB_genus'})
bacteriochlorophyll_genera

bacteriochlorophyll

# read genus statistics csv file and merge with the rhodopsin detection df. change NaN to No
Genera=pd.read_csv(outdir+'Supplemental_table_genus_statistics.csv', index_col=0)
Genera=Genera.merge(rhodopsin_genera, on='GTDB_genus', how='left')
Genera['Rhodopsin detected'].fillna('No', inplace=True)

# merge with the chlorophyll detection df. change NaN to No and write a new csv file
Genera=Genera.merge(bacteriochlorophyll_genera, on='GTDB_genus', how='left')
Genera['Chlorophyll detected'].fillna('No', inplace=True)
Genera.to_csv(outdir+'Supplemental_table_genus_statistics_light_metabolism.csv')
Genera

### Adding spherical cell volume



import math
Genera['spherical cell volume (µm3)']=(4/3)*math.pi*(Genera['median_estimated_diameter']/2)**3
Genera.to_csv(outdir+'Supplemental_table_genus_statistics_light_metabolism_volume.csv')
Genera


I need to add avg 16S copies per cell as well as avg mRNA transcripts per cell to this file. 

I should probably seperate it out by day so 3 or 4 columns for each.

rRNA=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/rRNA_recruitment/16S_rRNA_copies_per_cell_per_Genus.csv')
rRNA=rRNA.rename(columns={'Genus':'GTDB_genus'})
rRNA

#mRNA-pd.read_csv('')

df=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GORG_recruitment/RNA_summaries/Prokka_Transcripts_percell_O2_percell_Genus_stats_biogeochemical_master.csv',
              index_col=[1,2])

df = df.loc[:, :'Weighted_avg_O2_consumed_per_cell']
df=df.iloc[:, :-1]
df=df.drop('Unnamed: 0' , axis='columns')
df
mRNA=pd.DataFrame()
mRNA['total_transcripts']=df.sum(axis=1)
mRNA=mRNA.reset_index()
mRNA['date']=mRNA['date'].astype(str)
mRNA.rename(columns={'genus':'GTDB_genus'}, inplace=True)
mRNA


pivot=mRNA.pivot_table(index='GTDB_genus', columns='date', values='total_transcripts')

pivot.rename(columns={'20171102':'20171102_avg_mRNA_transcripts_per_cell', '20181030':'20181030_avg_mRNA_transcripts_per_cell',
                      '20190402':'20190402_avg_mRNA_transcripts_per_cell', '20190709':'20190709_avg_mRNA_transcripts_per_cell'}, inplace=True)

pivot=pivot.merge(rRNA, on='GTDB_genus', how='outer')
pivot

Genera=Genera.merge(pivot, on='GTDB_genus', how='left')
Genera.to_csv(outdir+'Supplemental_table_genus_statistics_light_metabolism_volume_RNA.csv')
Genera

! grep "dsr" /mnt/scgc/simon/microg2p/Data/SAG_genes/faa/GORG_Maine2020_noGEN.fixed.tsv > /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/tmp/dsrA.tsv
# cols=pd.read_csv('/mnt/scgc/simon/microg2p/Data/SAG_genes/faa/GORG_Maine2020_noGEN.fixed.tsv', sep='\t',  nrows=0).columns.tolist()
# bacteriochlorophyll=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/tmp/bacteriochlorophyll.tsv', sep='\t', names=cols)

# cols=['name', 'GTDB_classification']
# classification=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv', 
#                            usecols=cols)

# bacteriochlorophyll=bacteriochlorophyll.merge(classification, left_on='sag', right_on='name', how='left')

# # create a list of all genera where chlorophyll was detected
# bacteriochlorophyll_genera=bacteriochlorophyll[['GTDB_classification']].copy()
# bacteriochlorophyll_genera=bacteriochlorophyll_genera.drop_duplicates()
# bacteriochlorophyll_genera['Chlorophyll detected']='Yes'
# bacteriochlorophyll_genera=bacteriochlorophyll_genera.rename(columns={'GTDB_classification':'GTDB_genus'})
# bacteriochlorophyll_genera

# bacteriochlorophyll

! head /mnt/scgc/simon/microg2p/Data/SAG_genes/faa/GORG_Maine2020_noGEN.fixed.tsv
