==> 25_bacteroidetes_chl_analysis.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob
import seaborn as sns

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/chl_analysis/')

outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/chl_analysis/'
if os.path.isdir(outdir):
    print(outdir, "already exists.")
else:
    os.mkdir(outdir)
    print(outdir, "is created.")

images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/chl_analysis'
if os.path.isdir(images):
    print(images, "already exists.")
else:
    os.mkdir(images)
    print(images, "is created.")
    
matplotlib.__version__

rerun on 2021-12-07 after correcting for salinity of media samples



#SAG abundance. I need to convert this to cells/ml
SAG_abundance=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Syto9_SAG_abundances.csv',
                         index_col=0)

#key to only select bacteriodetes genera
classification=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Genus_statistics.csv',
                          index_col=0)
classification=classification.rename(columns={'GTDB_genus':'GTDB_classification'})
classification=classification[['Phyla', 'Family', 'GTDB_classification']]

# chlorophyl concentration
chl_concentration=pd.read_csv('G2P_DS_physical_chl_extract_20181030_20200616.csv')


classification


# the total number of cells/ml from Syto9 stains.
cells_1704=1545111
cells_1708=2690141
cells_1711=1550260
cells_1810=1452500
cells_1904=1642105
cells_1907=2915238

SAG_abundance=SAG_abundance.merge(classification, on='GTDB_classification', how='left')
Bacteriodetes=SAG_abundance[SAG_abundance['Phyla']=='Bacteroidota']

Bacteriodetes['1704_cells']=Bacteriodetes['170412_SAG_abundnance']*cells_1704
Bacteriodetes['1708_cells']=Bacteriodetes['170818_SAG_abundnance']*cells_1708
Bacteriodetes['1711_cells']=Bacteriodetes['171102_SAG_abundnance']*cells_1711
Bacteriodetes['1810_cells']=Bacteriodetes['181030_SAG_abundnance']*cells_1810
Bacteriodetes['1904_cells']=Bacteriodetes['190402_SAG_abundnance']*cells_1904
Bacteriodetes['1907_cells']=Bacteriodetes['190709_SAG_abundnance']*cells_1907

#sum up all the bacterodiota cells and put into a series
bacter_1704=Bacteriodetes['1704_cells'].sum()
bacter_1708=Bacteriodetes['1708_cells'].sum()
bacter_1711=Bacteriodetes['1711_cells'].sum()
bacter_1810=Bacteriodetes['1810_cells'].sum()
bacter_1904=Bacteriodetes['1904_cells'].sum()
bacter_1907=Bacteriodetes['1907_cells'].sum()
counts=[bacter_1704, bacter_1708, bacter_1711, bacter_1810, bacter_1904, bacter_1907]
counts

bacteriodia=pd.DataFrame()
bacteriodia['date']=[20170412, 20170818, 20171102, 20181030, 20190402, 20190709]
bacteriodia['Bacteroidia_SAGs_/ml']=pd.Series(counts)
bacteriodia['total_cells/ml']=[1545111, 2690141, 1550260, 1452500, 1642105, 2915238]

bacteriodia['%bacteriodia_SAGs']=(bacteriodia['Bacteroidia_SAGs_/ml']/bacteriodia['total_cells/ml'])*100

bacteriodia

#metagenome based genus abundance
mg_abundance=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/GoM_Metagenome_analysis/GoM_metagenome_calculated_Genus_cells_per_ml.csv',
                        index_col=0)


mg_abundance=mg_abundance.merge(classification, left_on='genus', right_on='GTDB_classification', how='left')
mg_Bacteroidiota=mg_abundance[mg_abundance['Phyla']=='Bacteroidota']

mg_Bacteroidiota.drop(mg_Bacteroidiota.columns[[0,1,2,3,9,10,11]], axis=1, inplace=True)

mg_Bacteroidiota

mg_1711=mg_Bacteroidiota['20171102_genome_size_normalized_metagenome_cells/ml'].sum()
mg_1810=mg_Bacteroidiota['20181030_genome_size_normalized_metagenome_cells/ml'].sum()
mg_1904=mg_Bacteroidiota['20190402_genome_size_normalized_metagenome_cells/ml'].sum()
mg_1907=mg_Bacteroidiota['20190709_genome_size_normalized_metagenome_cells/ml'].sum()
mg_counts=[mg_1711, mg_1810, mg_1904, mg_1907]

mg_bacteriodia=pd.DataFrame()
mg_bacteriodia['date']=[20171102, 20181030, 20190402, 20190709]
mg_bacteriodia['mg_bacteroidia_cells/ml']=pd.Series(mg_counts)
mg_bacteriodia

bacteriodia=bacteriodia.merge(mg_bacteriodia, on='date', how='left')
bacteriodia['%mg_bacteroidia_cells/ml']=(bacteriodia['mg_bacteroidia_cells/ml']/bacteriodia['total_cells/ml'])*100
bacteriodia


chl_concentration['SampleDate']=chl_concentration['SampleDate'].str.replace("-", '')
chl=chl_concentration[['SampleDate', 'ChlWWug/L', 'SampleLocale']].copy()
chl=chl.rename(columns={'SampleDate':'date'})
chl

bacteriodia['date']=bacteriodia['date'].astype(str)
bacteriodia=bacteriodia.merge(chl, on='date', how='outer')
bacteriodia['SampleLocale'].fillna('ebb', inplace=True)
bacteriodia.sort_values(by='date', inplace=True)

bacteriodia=bacteriodia.dropna(axis=0, thresh=3)
bacteriodia=bacteriodia.sort_values(by='date')
bacteriodia

df=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/chl_analysis/DS_chl_measurements.csv')

df[['month', 'day', 'year']] = df.DATE.str.split('/', expand = True)
df['year']=pd.to_numeric(df['year'])+2000
df['year']=df['year'].astype(str)
# add a zero befor the days and months that are a single digit
df['day'] = df['day'].apply(lambda x: x.zfill(2))
df['month'] = df['month'].apply(lambda x: x.zfill(2))

df['date']=df['year']+df['month']+df['day']

df=df.rename(columns={'WW Mean':'ChlWWug/L'})
df=df.drop(columns=['DATE', 'month', 'day', 'year'])
df
#print(df.dtypes)

chl=pd.concat([bacteriodia, df])
chl['year']=chl['date'].str[:4]
chl['month']=chl['date'].str[5:6]
chl['day']=chl['date'].str[6:]
chl['date']=chl['date'].astype(int)
chl=chl[chl['date']>=20170000]
chl=chl[chl['date']<=20200000]
chl['SampleLocale'].fillna('bbh', inplace=True)


chl=chl.sort_values(by='date')
chl.to_csv(outdir+'2017_2019_chla_measurements.csv')
chl=chl.drop(columns=['total_cells/ml',  'year', 'month', 'day'])
chl=chl.rename(columns={'Bacteroidia_SAGs_/ml':'SAG_per_ml', 'mg_bacteroidia_cells/ml': 'mg_per_ml', 'ChlWWug/L':'ChlWW',
                        '%mg_bacteroidia_cells/ml':'perc_mg_bacteroidia', '%bacteriodia_SAGs':'perc_SAGs_bacteroidia'})
chl.to_csv(outdir+'2017_2019_chla_measurements.tsv', sep = '\t', index=False)
chl

### I might need to add a column where I calculate the number of RSG bacteriodeta cells to supplement this analysis.

Flavobacteraceae=SAG_abundance[SAG_abundance['Family']=='Flavobacteriaceae']

Flavobacteraceae['1704_cells']=Flavobacteraceae['170412_SAG_abundnance']*cells_1704
Flavobacteraceae['1708_cells']=Flavobacteraceae['170818_SAG_abundnance']*cells_1708
Flavobacteraceae['1711_cells']=Flavobacteraceae['171102_SAG_abundnance']*cells_1711
Flavobacteraceae['1810_cells']=Flavobacteraceae['181030_SAG_abundnance']*cells_1810
Flavobacteraceae['1904_cells']=Flavobacteraceae['190402_SAG_abundnance']*cells_1904
Flavobacteraceae['1907_cells']=Flavobacteraceae['190709_SAG_abundnance']*cells_1907

#sum up all the bacterodiota cells and put into a series
flavo_1704=Flavobacteraceae['1704_cells'].sum()
flavo_1708=Flavobacteraceae['1708_cells'].sum()
flavo_1711=Flavobacteraceae['1711_cells'].sum()
flavo_1810=Flavobacteraceae['1810_cells'].sum()
flavo_1904=Flavobacteraceae['1904_cells'].sum()
flavo_1907=Flavobacteraceae['1907_cells'].sum()
counts=[flavo_1704, flavo_1708, flavo_1711, flavo_1810, flavo_1904, flavo_1907]
counts

Flavo=pd.DataFrame()
Flavo['date']=[20170412, 20170818, 20171102, 20181030, 20190402, 20190709]
Flavo['SAG_cells/ml']=pd.Series(counts)
Flavo['total_cells/ml']=[1545111, 2690141, 1550260, 1452500, 1642105, 2915238]

Flavo['%Flavo']=(Flavo['SAG_cells/ml']/Flavo['total_cells/ml'])*100

Flavo
