==> 26b_Surface_ocean_analysis_combination.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob
import seaborn as sns


outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis/'
if os.path.isdir(outdir):
    print(outdir, "already exists.")
else:
    os.mkdir(outdir)
    print(outdir, "is created.")

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis/')    

images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/Surface_ocean_SAG_analysis'
if os.path.isdir(images):
    print(images, "already exists.")
else:
    os.mkdir(images)
    print(images, "is created.")
    
matplotlib.__version__

rerun on 2021-12-07 after correcting for salinity of media samples

Sintes_SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis/Sintes_SAGs_GTDB_master_predresp_rate.csv',
                       index_col=0)

Sintes_SAGs

JdF_SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/JdF_analysis/summary_files/JdF_SAGs_GTDB_1.5_and_95ANI_clustering_master_predresp_rate.csv',
                    index_col=0)
JdF_SAGs
JdF_Surf_plates=['AH-779', 'AH-883']
JdF_Surf=JdF_SAGs[JdF_SAGs['Plate'].isin(JdF_Surf_plates)]

JdF_SAGs=JdF_SAGs[~JdF_SAGs.name.str.contains("-H")]
JdF_Surf

test=JdF_SAGs[JdF_SAGs['name']=="AH-779-F04"]
test

Surface_SAGs=pd.concat([Sintes_SAGs, JdF_Surf])

Surface_SAGs.to_csv(outdir+'Surface_SAGs_GTDB_master_predresp_rate.csv')

with open('Surface_SAGs.txt', 'w') as f:
    for text in Surface_SAGs['name'].tolist():
        f.write(text+'\n')

Surface_SAGs



I manually copied the SSU sequences into the folder below because they were in multiple places and i did not want to spend the time figuring out how to code it to look in 5 different folders to copy the SSU sequences over.


SSU='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis/Surface_SSU_seqs'
if os.path.isdir(SSU):
    print(SSU, "already exists.")
else:
    os.mkdir(SSU)
    print(images, "is created.")
    

Modify the code below so that it creates a fasta file with the longest sequence for every genus I detected in these surface samples. I am going to have to change it since I do not have this genus summary file created. (at least not yet).

Genus=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis/Surface_SAGs_GTDB_master_predresp_rate.csv', index_col=0)
Genus=Genus[Genus['gate'] == 'Respiring cells (RSG-positive)']
Genus=Genus[Genus['GTDB_classification'] != 'Unclassified']
tenSAG=Genus['GTDB_classification'].tolist()
tenSAG=np.unique(tenSAG)
print(len(tenSAG))

SAG_data=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis/Surface_SAGs_GTDB_master_predresp_rate.csv',
                    index_col=0)
SAG_data

# remove SAGs that have been identified as contaminated via tetramer analysis 
# find all the SSU fasta files, grep the file names using a txt file, delete the ones with a partial match in the txt file
!find Surface_SSU_seqs/* | grep -f /mnt/scgc/simon/microg2p/Data/Lists/contaminated_SAGs.txt | xargs rm

#create a fasta file for every genus
outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis/Surface_SSU_seqs/'
!rm *.fasta
SSU_dir=glob.glob('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis/Surface_SSU_seqs/*.fasta')

for genus in tenSAG:
    df=SAG_data[SAG_data['GTDB_classification'] == genus]
    SAGs=df['name'].to_list()

    with open(genus+'_SSU_sequences.fa', 'w') as out:
        for i in SAGs:
            for infile in SSU_dir:
                if i in infile:
                    with open(infile, 'r') as infile:
                        out.write(infile.read())
                        
!mv *.fa Surface_SAGs_SSU_by_genus/


os.chdir('Surface_SAGs_SSU_by_genus/')
!pwd
# sort the 16S sequences long to short
!for file in *_SSU_sequences.fa; do seqkit sort $file -o ${file/%SSU_sequences.fa/sorted.fa} -l -r; done

# extract just the longest 16S sequence
!for file in *sorted.fa; do awk '/^>/{if(N)exit;++N;} {print;}' $file > ${file/%sorted.fa/longest16S.fa}; done

#remove the sorted files
!rm *sorted.fa

# rename the name of the longest sequence so that it is the name of the genus
# This helps with future annotations of a tree where I have to refer to the branch length
fasta_files=glob.glob('*16S.fa')

for i in fasta_files:
    
    genus=i.split("_lon")[0]
    genus=str(genus)
   # print(genus)
    fasta_in = open(i)

    # Create FASTA output file.
    fasta_out = open(genus+'_16Srep.fa', 'w')

    for line in fasta_in.readlines():
        if line.startswith('>'):
            contig_id = '>' + str(genus) + '\n'
 #           print(contig_id)
            fasta_out.write(contig_id)
        else:
            fasta_out.write(line)
    fasta_out.close()
    fasta_in.close()


# combine all the longest representatives into a single fasta file.
os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis')
!pwd
!rm Genus_longest_16S_sequence.fa
!rm *longest16S.fa
!cat Surface_SAGs_SSU_by_genus/*16Srep.fa >> Genus_longest_16S_sequence.fa

!seqkit stats Genus_longest_16S_sequence.fa


# create a file of all 16S sequence for each day


os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/Surface_ocean_SAG_analysis')
!cat Surface_SSU_seqs/AH-558* > AH-558_SSU_sequences.fasta
!cat Surface_SSU_seqs/AH-736* > AH-736_SSU_sequences.fasta
!cat Surface_SSU_seqs/AH-716* > AH-716_SSU_sequences.fasta
!cat Surface_SSU_seqs/AH-779* > AH-779_SSU_sequences.fasta
!cat Surface_SSU_seqs/AH-883* > AH-883_SSU_sequences.fasta

!for file in *_SSU_sequences.fasta; do seqkit seq $file -m 1200 -o ${file/%SSU_sequences.fasta/SSU_1200bp_sequences.fasta} -l -r; done
!rm *SSU_sequences.fasta


Surface_O2=SAG_data[['name', 'Predicted_respiration_rate_fmol_O2perhr', 'gate', 'Plate', 'GTDB_classification', 'estimated_diameter']]
Surface_O2=Surface_O2[Surface_O2['gate']=='Respiring cells (RSG-positive)']
unclass=Surface_O2.shape[0]
print('The number of SAGs including unclassified is '+ str(unclass))

#Surface_O2=Surface_O2[Surface_O2['GTDB_classification'] != 'Unclassified']


d={'Plate':['AH-558', 'AH-716', 'AH-736', 'AH-779'], 'depth':[100,5,5,10]}
depth=pd.DataFrame(data=d)
Surface_O2=Surface_O2.merge(depth, on='Plate', how='left')

Surface_O2.to_csv('Surface_SAGs_O2_consumption_rates_for_trees.csv')
Surface_O2
