==> 27_apr_analysis.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob
import seaborn as sns


outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/aprA/'
if os.path.isdir(outdir):
    print(outdir, "already exists.")
else:
    os.mkdir(outdir)
    print(outdir, "is created.")

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/aprA/')    

images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/aprA'
if os.path.isdir(images):
    print(images, "already exists.")
else:
    os.mkdir(images)
    print(images, "is created.")
    
matplotlib.__version__

rerun on 2021-12-07 after correcting for salinity of media samples

!module use /mod/scgc
!module load seqtk/1.0.75




! grep "Adenylylsulfate" /mnt/scgc/simon/microg2p/Data/SAG_genes/all_GoM_genes.ffn > apr_genes.txt
! grep "alpha" apr_genes.txt > aprA_genes.txt
alpha=pd.read_csv('aprA_genes.txt', sep=" ", header=None, usecols=[0])
alpha[0]=alpha[0].str[1:]
alist=list(alpha[0])

file=open('aprA_genes.txt','w')
for items in alist:
    file.writelines(items+'\n')
file.close()


! grep "beta" apr_genes.txt > aprB_genes.txt
beta=pd.read_csv('aprB_genes.txt', sep=" ", header=None, usecols=[0])
beta[0]=beta[0].str[1:]
blist=list(beta[0])

file=open('aprB_genes.txt','w')
for items in blist:
    file.writelines(items+'\n')
file.close()

! head aprA_genes.txt

!seqkit grep -f aprA_genes.txt /mnt/scgc/simon/microg2p/Data/SAG_genes/all_GoM_genes.ffn  -o aprA_genes.fasta
!seqkit grep -f aprB_genes.txt /mnt/scgc/simon/microg2p/Data/SAG_genes/all_GoM_genes.ffn  -o aprB_genes.fasta


! head aprA_genes.fasta

! seqkit translate aprA_genes.fasta -o aprA_genes.faa --trim
! seqkit translate aprB_genes.fasta -o aprB_genes.faa --trim
! head aprB_genes.faa

SAG_data=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv', 
                     index_col=0, usecols=['name', 'GTDB_classification'])
SAG_data
alpha[0]=alpha[0].str[:10]
alpha=alpha.merge(SAG_data, how='left', left_on=0, right_on='name')
alpha=alpha.rename(columns={0:'SAG'})

beta[0]=beta[0].str[:10]
beta=beta.merge(SAG_data, how='left', left_on=0, right_on='name')
beta=beta.rename(columns={0:"SAG"})

alpha

#check the length of the genes and get rid of any that are <75% of full length
!seqkit stats aprA_genes.faa
!seqkit stats aprB_genes.faa

# remove any sequences that are <75% of the average length 
!seqkit seq -m 450 aprA_genes.faa > aprA_genes_75perc.faa
!seqkit seq -m 120 aprB_genes.faa > aprB_genes_75perc.faa

! cat aprA_genes_75perc.faa aprA_ref_genes.faa > aprA_genes_75perc_and_ref.faa

!seqkit stats aprA_genes_75perc.faa
!seqkit stats aprB_genes_75perc.faa

! module load muscle/3.8.31
! muscle -in aprA_genes_75perc_and_ref.faa.faa -out aprA_genes_75perc_and_ref_muscle_align.faa
! muscle -in aprB_genes_75perc.faa -out aprB_75perc_muscle_align.faa
