==> 29_GoM_202108_GoM_resp_Summary_file_creator_and_decon.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob
import seaborn as sns

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/202108_GoM_resp/Assembly_and_FACS_files/')

outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files'
if os.path.isdir(outdir):
    print(outdir, "already exists.")
else:
    os.mkdir(outdir)
    print(outdir, "is created.")

images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/202108_GoM_resp_expts'
if os.path.isdir(images):
    print(images, "already exists.")
else:
    os.mkdir(images)
    print(images, "is created.")
    
matplotlib.__version__

AM_FACS=pd.DataFrame()
AM_SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/GoM_202108_SAGs_GTDBclass.csv', index_col=0)

read_files = glob.glob('*.csv')
for f in read_files:
    df=pd.read_csv(f, index_col=0)
    plate=f[:7]
    df['well']=(plate+df['well'])
    AM_FACS=AM_FACS.append(df)
    print (plate)
AM_SAGs=AM_SAGs.merge(AM_FACS, on='well')

AM_SAGs=AM_SAGs.rename(columns={'well':'name'})

AM_SAGs



os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/202108_GoM_resp/')

AM_SAGs=AM_SAGs.reset_index()
AM_SAGs.drop(['index'], axis=1, inplace=True)

AM_SAGs=AM_SAGs[AM_SAGs['well_type']=='1 cell']
#AM_SAGs=AM_SAGs.rename(columns={'name':'well', 'well':'name'})
AM_SAGs['Plate']=AM_SAGs['name'].str[:6]
AM_SAGs['well']=AM_SAGs['name'].str[7:]

# I need to add a column that is gate and fill it as RSG for 775, and 777
Syto=['AM-382', 'AM-384']
AM_SAGs['gate']=np.where(AM_SAGs.Plate.isin(Syto),'All prokaryotes (SYTO-9-positive)', 'Respiring cells (RSG-positive)')

AM_SAGs.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/202108_GoM_resp_all_1cell_SAGs.csv')
AM_SAGs



# I think that the next step is that I need to look over the tetramer PCA analysis and remove any that look like they are contaminated
with open("/mnt/scgc/simon/microg2p/Data/Lists/contaminated_SAGs.txt") as file:
    bad_SAGs = [line.strip() for line in file]
    
AM_SAGs=AM_SAGs[~AM_SAGs.name.isin(bad_SAGs)]
AM_SAGs.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/202108_GoM_resp_decon_1cell_SAGs.csv')
AM_SAGs

# Add the date and depth to the samples
AM_SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/202108_GoM_resp_decon_1cell_SAGs.csv', index_col=0)

Tues=['AM-377', 'AM-384']
AM_SAGs['sample_collection_date']=np.where(AM_SAGs.Plate.isin(Tues), 210824, 210825)
AM_SAGs['sample_collection_time']=np.where(AM_SAGs.Plate.isin(Tues), 'Day', 'Night')

AM_SAGs['FACS date']='210928'

AM_SAGs['depth']=1

# add normalization values for RSG
cols=['date', 'slope (m)', 'intercept (b)']
RSG_conversion=pd.read_csv('531_conversion_factors.csv', usecols=cols)
RSG_conversion=RSG_conversion.rename(columns={'date':'FACS date'})
RSG_conversion['FACS date']=RSG_conversion['FACS date'].astype(str)
AM_SAGs=AM_SAGs.merge(RSG_conversion, on='FACS date', how='left')


AM_SAGs['normalized_531_value']=np.where(AM_SAGs['gate']=='Respiring cells (RSG-positive)', ((AM_SAGs['X531.40..488.']*AM_SAGs['slope (m)'])+AM_SAGs['intercept (b)']), '')

AM_SAGs.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/202108_GoM_resp_decon_1cell_SAGs_date_normalized_RSG.csv')
AM_SAGs

# The rows without a normalized 531 are all Syto9 stained SAGs and do not need to be converted.
