==> 32_GoM_resp_dotplot_generator.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob
import seaborn as sns

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/202108_GoM_resp/')

outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/202108_GoM_resp/'
if os.path.isdir(outdir):
    print(outdir, "already exists.")
else:
    os.mkdir(outdir)
    print(outdir, "is created.")

images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/202108_GoM_resp_expts'
if os.path.isdir(images):
    print(images, "already exists.")
else:
    os.mkdir(images)
    print(images, "is created.")
    
matplotlib.__version__

cols=['name', 'GTDBtk_classification', 'Plate',	'gate']

SAG_df=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/202108_GoM_resp_decon_1cell_SAGs_master_predresp_rate.csv',
                   usecols=cols)

SAG_df=SAG_df[SAG_df['gate']=='Respiring cells (RSG-positive)']
SAG_df[['domain', 'phyla', 'class', 'order', 'family', 'genus', 'species']]=SAG_df.GTDBtk_classification.str.split(';', expand=True)

SAG_df=SAG_df.drop(['domain', 'phyla', 'order', 'family', 'genus', 'species'] , axis='columns')
SAG_df['class']=SAG_df['class'].str[3:]
SAG_df=SAG_df.drop(['GTDBtk_classification', 'gate'], axis='columns')
SAG_df['class'].fillna(value='Other')

AM375=SAG_df[SAG_df['Plate']=='AM-375']
AM375=AM375.drop('Plate', axis='columns')
AM375.to_csv('AM375_dotplot.csv', index=False, header=False)


AM377=SAG_df[SAG_df['Plate']=='AM-377']
AM377=AM377.drop('Plate', axis='columns')
AM377.to_csv('AM377_dotplot.csv', index=False, header=False)
AM377

!cat FACS_header.csv AM377_dotplot.csv > Final_AM_377_dotplot.csv

!cat FACS_header.csv AM375_dotplot.csv > Final_AM_375_dotplot.csv



!cp Final*.csv /mnt/scgc_raw/clarity/dropoff/facs_labeled_plot
