==> 34a_Cryopreservation_table_munging.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob
import seaborn as sns
pd.set_option("display.max_rows", 10)


outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/CRYO_analysis'
if os.path.isdir(outdir):
    print(outdir, "already exists.")
else:
    os.mkdir(outdir)
    print(outdir, "is created.")

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/CRYO_analysis')    

images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/CRYO_analysis'
if os.path.isdir(images):
    print(images, "already exists.")
else:
    os.mkdir(images)
    print(images, "is created.")
    
matplotlib.__version__


SAG_data=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_5SAG_genera_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv', index_col=0)




cryo_data=SAG_data[SAG_data['Plate'].isin(['AH-545', 'AH-580'])]
cryo_data.drop(columns='X2', inplace=True)
cryo_data.columns


new_plate=pd.read_csv('tbls/211209_NB502014_0167_AHK5HTBGXK_assembly_stats.csv')
new_plate.columns

col_order=['name', 'Plate', 'well_type', 'wga_cp', 'wgs_run_id', 'raw_read_count',
       'final_assembly_length', 'max_contig_length', 'gc_content',
       'checkM_estimated_completeness', 'total_CDS_count',
       'hypothetical_cds_fraction', 'average_cds_length', 'number_multi_copy',
       'checkM_classification', 'SSU_classification_1', 'SSU_classification_2',
       'SSU_classification_3', 'notes']


new_plate=new_plate.rename(columns={'well':'name'})
new_plate['Plate']=new_plate['name'].str[:6]
new_plate['well_type']='1 cell'
new_plate['wga_cp']=''
new_plate['average_cds_length']=''
new_plate['SSU_classification_3']=''
new_plate['hypothetical_cds_fraction']=''
new_plate['SSU_classification_2']=''
new_plate['total_CDS_count']=''


new_plate=new_plate[col_order]
new_plate

# add FACS data
col_order=['name', 'Plate', 'well_type', 'wga_cp', 'wgs_run_id', 'raw_read_count',
       'final_assembly_length', 'max_contig_length', 'gc_content',
       'checkM_estimated_completeness', 'total_CDS_count',
       'hypothetical_cds_fraction', 'average_cds_length', 'number_multi_copy',
       'checkM_classification', 'SSU_classification_1', 'SSU_classification_2',
       'SSU_classification_3', 'notes', 'eventnum', 'well2', 'X531.40..488.',
       'X572.27..488.', 'X692.40..488.', 'Trigger.Pulse.Width', 'SSC', 'FSC',
       'estimated_diameter']

# open the FACS data I do not see the csv file for this sample from Nicole
FACS_data=pd.read_csv('/mnt/scgc/simon/microg2p/Data/SAG_FACS/FACS_csv_files/AM-379-data-export.csv')

FACS_data['well']='AM-379-'+FACS_data['well']
FACS_data.rename(columns={'well':'well2'}, inplace=True)
# FACS_data
new_plate=new_plate.merge(FACS_data, how='left', left_on='name', right_on='well2')
new_plate


new_plate=new_plate[col_order]

# add other metadata
col_order=['name', 'Plate', 'well_type', 'wga_cp', 'wgs_run_id', 'raw_read_count',
       'final_assembly_length', 'max_contig_length', 'gc_content',
       'checkM_estimated_completeness', 'total_CDS_count',
       'hypothetical_cds_fraction', 'average_cds_length', 'number_multi_copy',
       'checkM_classification', 'SSU_classification_1', 'SSU_classification_2',
       'SSU_classification_3', 'notes', 'eventnum', 'well2', 'X531.40..488.',
       'X572.27..488.', 'X692.40..488.', 'Trigger.Pulse.Width', 'SSC', 'FSC',
       'estimated_diameter', 'PCA_analysis_Action', 'PCA_analysis_Note',
       'Location', 'Sample_collection_Date', 'FACS_date', 'gate', 'depth..m.',
       'sample', 'slope', 'intercept', 'normalized_531_value']

new_plate['PCA_analysis_Action']=np.nan
new_plate['PCA_analysis_Note']=np.nan
new_plate['Location']='EBB Bigelow dock'
new_plate['Sample_collection_Date']=181030
new_plate['FACS_date']=210928
new_plate['gate']='Respiring cells (RSG-positive)'
new_plate['depth..m.']=1
new_plate['sample']=new_plate['name']




# add RSG normalization values and normalize 531 data for O2 consumption calculation
cols=['date', 'slope (m)', 'intercept (b)']
RSG_conversion=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/202108_GoM_resp/531_conversion_factors.csv', usecols=cols)
RSG_conversion=RSG_conversion.rename(columns={'date':'FACS_date', 'slope (m)':'slope', 'intercept (b)':'intercept'})
#RSG_conversion['FACS date']=RSG_conversion['FACS date'].astype(str)
new_plate=new_plate.merge(RSG_conversion, on='FACS_date', how='left')


new_plate['normalized_531_value']=np.where(new_plate['gate']=='Respiring cells (RSG-positive)', ((new_plate['X531.40..488.']*new_plate['slope'])+new_plate['intercept']), '')

new_plate=new_plate[col_order]

new_plate.to_csv(os.path.join(outdir, "tbls/AM-389_Assembly_stats_and_metadata.csv"))

new_plate
