==> 34c_final_Cryo_table_munging.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob
import seaborn as sns
pd.set_option("display.max_rows", 10)


outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/CRYO_analysis'
if os.path.isdir(outdir):
    print(outdir, "already exists.")
else:
    os.mkdir(outdir)
    print(outdir, "is created.")

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/CRYO_analysis')    

images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/CRYO_analysis'
if os.path.isdir(images):
    print(images, "already exists.")
else:
    os.mkdir(images)
    print(images, "is created.")
    
matplotlib.__version__

SAG_data=pd.read_csv('tbls/AM-379_predicted_O2_respiration_rates.csv', index_col=0)
SAG_data

arc_GTDB_class=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/CRYO_analysis/cryo_RSG_gtdb_1.5_classify_wf_out/classify/gtdbtk.ar122.summary.tsv', sep='\t')

bac_GTDB_class=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/CRYO_analysis/cryo_RSG_gtdb_1.5_classify_wf_out/classify/gtdbtk.bac120.summary.tsv', sep='\t')
GTDB_class=pd.concat([arc_GTDB_class, bac_GTDB_class])
GTDB_class.drop_duplicates(inplace=True)
GTDB_class['SAG']=GTDB_class['user_genome'].str[:10]

GTDB_class

classification=GTDB_class[['user_genome', 'classification']].copy()
classification[['domain', 'phyla', 'class', 'order', 'family', 'genus', 'species']]=classification.classification.str.split(';', expand=True)
classification=classification.rename(columns = {'classification':'GTDBtk_classification'})
classification['count']=1
classification.to_csv('tbls/AM-379_gtdb_classify_wf_classifications_ar122_bac120.csv')
classification

### Big Question 
How many of the genera have more than 1 species present? I can not count the unassigned species but I can look at those with multiple assigned species.

# classify down to the Genus level and use this as my classification

classification=pd.read_csv('tbls/AM-379_gtdb_classify_wf_classifications_ar122_bac120.csv')
classification['phyla']=classification['phyla'].str[3:]
classification['class']=classification['class'].str[3:]
classification['order']=classification['order'].str[3:]
classification['family']=classification['family'].str[3:]
classification['genus']=classification['genus'].str[3:]
classification['species']=classification['species'].str[3:]
classification


classification['GTDB_classification'] = classification['phyla']

classification.loc[classification['class'] != '', 'GTDB_classification'] = classification['class']
classification.loc[classification['order'] != '', 'GTDB_classification'] = classification['order']
classification.loc[classification['family'] != '', 'GTDB_classification'] = classification['family']
classification.loc[classification['genus'] != '', 'GTDB_classification'] = classification['genus']

classification=classification.rename(columns={'user_genome':'SAG'})
classification.to_csv('202108_GoM_gtdb_classify_wf_classifications.csv')

classification

GTDB_class=classification[['SAG', 'GTDB_classification', 'GTDBtk_classification']]
GTDB_class['SAG']=GTDB_class['SAG'].str[:10]
GTDB_class

SAG_data=SAG_data.merge(GTDB_class, how='left', left_on='name', right_on='SAG')
# SAG_data.drop(columns={'SAG'})

SAG_data['GTDB_classification']=SAG_data['GTDB_classification'].fillna('Unclassified')

SAG_data['GTDB_classification'].value_counts()

SAG_data.to_csv(os.path.join(outdir, 'AM-379_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv'))

SAG_data

All_SAG_data=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_5SAG_genera_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv', index_col=0)
cryo_data=All_SAG_data[All_SAG_data['Plate'].isin(['AH-545', 'AH-580'])]
cryo_data.drop(columns='X2', inplace=True)
cryo_data

cryo_data=pd.concat([cryo_data, SAG_data])
cryo_data

cols=['name', 'GTDB_classification', 'Predicted_respiration_rate_fmol_O2perhr', 'FACS_date', 
      'Plate', 'GTDBtk_classification']

heatmap_data=cryo_data[cols].copy()
heatmap_data.to_csv('tbls/data_for_heatmap_analysis.csv')
heatmap_data

# ADD the section to generate dotplots but make sure that I modify the values so that I am using the same colors and albels as the rest of the paper.
Plates=heatmap_data.Plate.unique()

    
cols=['name', 'GTDBtk_classification']
dotplot=heatmap_data[cols].copy()

# I am doing this one step at a time because if I loop it I am worried that I will lose the 
# ability to control the order in which each lineage is tested. Eg Bacteria will be last and all the 
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('a'), 'Other', '')
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Alphaproteobacteria'), 'Alphaproteobacteria', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Planctomycetes'), 'Planctomycetes', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Gammaproteobacteria'), 'Gammaproteobacteria', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Poseidoniia'), 'Poseidoniia', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Nitrososphaeria'), 'Nitrososphaeria', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Verrucomicrobiae'), 'Verrucomicrobiae', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Acidimicrobiia'), 'Acidimicrobiia', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Actinomycetia'), 'Actinomycetia', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Marinisomatia'), 'Marinisomatia', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Kapabacteria'), 'Kapabacteria', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Bacteroidia'), 'Bacteroidia)', dotplot['dot_classification'])

dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Cyanobacteriia'), 'Cyanobacteriia', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('UBA1135'), 'UBA1135', dotplot['dot_classification'])
dotplot['dot_classification']=np.where(dotplot.GTDBtk_classification.str.contains('Campylobacteria'), 'Campylobacteria', dotplot['dot_classification'])

outdir='./dotplot_tbls/'
if not os.path.exists(outdir):
    os.mkdir(outdir)
    
! rm dotplot_tbls/A*.csv

for p in Plates:
    tdf=dotplot[dotplot.name.str.contains(p, na=False)]
    tdf=tdf.drop(columns=['GTDBtk_classification'])
    tdf.to_csv(outdir+ p+'_dotplot.csv', index=False, header=False)
    



# combine with the header fle that I have manually generated
! for file in ./dotplot_tbls/A*plot.csv; do cat /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/CRYO_analysis/dotplot_tbls/dotplot_header.csv $file > ${file/%.csv/_final.csv}; done

! cp ./dotplot_tbls/A*final.csv /mnt/scgc_raw/clarity/dropoff/facs_labeled_plot

! rm files_for_dotplots.zip
# zip final files
!tar -czvf files_for_dotplots.zip ./dotplot_tbls/*final.csv 

# copy to the facs_labeled_plot folder to generate dotplots
!cp files_for_dotplots.zip /mnt/scgc_raw/clarity/dropoff/facs_labeled_plot
