==> 35_DRAM_cleanup.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob
import seaborn as sns
from collections import Counter
pd.set_option("display.max_rows", 10)

matplotlib.__version__



os.chdir('/mnt/scgc/simon/microg2p/Data/Decoupling_respiration_rates_contigs')

SAG_df=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                  index_col=0)

new_SAG_df=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/202108_GoM_resp_decon_1cell_SAGs_master_predresp_rate.csv',
                      index_col=0)


SAG_df.columns

all_SAG_df=pd.concat([SAG_df,new_SAG_df], axis=0, ignore_index=True)
all_SAG_df

# I need to cleanup these dataframes so that they ahve the same columns and then generate a master file


clean_SAGs=SAG_df['name'].to_list()
clean_new_SAGs=new_SAG_df['name'].to_list()
all_clean_SAGs=clean_SAGs+clean_new_SAGs
print(len(clean_SAGs))
print(len(clean_new_SAGs))

len(all_clean_SAGs)

with open('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Decoupling_respiration_rates_clean_SAGs.txt', 'w') as f:
    for item in all_clean_SAGs:
        f.write("%s\n" % item)


! find . -name '*contigs.fasta' | grep -v -f  /mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Decoupling_respiration_rates_clean_SAGs.txt | xargs rm


len(all_clean_SAGs)
