==> 97_files_for_figures.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob

# input folder
os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/gtdb')

# folder to save output files will create it if it does not exist
outdir="/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/csv_files_for_figures/"
if os.path.isdir(outdir):
    print(outdir, "already exists.")
else:
    os.mkdir(outdir)
    print(outdir, "is created.")

matplotlib.__version__

rerun on 2021-12-07 after correcting for salinity of media samples

## Figure 2 csv files to create

Upload the color key file

1. abundance of all genera
2. O2 consumption of all SAGs labeled with RSG as well as their genus
3. Total O2 consumption /ml for all genera

with open("/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Genera_with_5+SAGs.txt", "r") as tf:
    five_SAG = tf.read().split('\n')
print(five_SAG)



# abundance per day and weighted average
a_df=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/avg_O2_consumption_and_abundance_per_day_melt.csv',
                index_col=0)
#a_df=a_df[a_df['genus'].isin(five_SAG)]
print(a_df['genus'].nunique())
a_df

#Parse df down to just the genera with at least 5 SAGs

abundance_df=a_df[['genus', 'date', 'cells_per_ml']].copy()
abundance_df.to_csv(outdir+'Fig2_abundance.csv')

# CSV for boxplot 
cols=['name', 'Predicted_respiration_rate_fmol_O2perhr', 'GTDB_classification', 'gate', 'FACS_date']
SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                usecols=cols)

# per cell respiration rates
SAGs=SAGs[SAGs['gate']=='Respiring cells (RSG-positive)']
dates=[181030, 190402, 190709]
SAGs=SAGs[SAGs['FACS_date'].isin(dates)]
#SAGs=SAGs[SAGs['GTDB_classification'].isin(five_SAG)]
SAGs.to_csv(outdir+'per_cell_O2_consumption_rates_for_fiveSAG_genera.csv')

#average_weighted_O2 conusmption rates per genus
avg_weighted_df=a_df[['genus', 'date', 'avg_fmol_O2_consumed']]
avg_weighted_df.to_csv(outdir+'average_weighted_O2_consumption_per_cell.csv')
SAGs

# total O2 conusmption per genus.
a_df['Total_O2_consumption']=a_df['cells_per_ml']*a_df['avg_fmol_O2_consumed']
perml_O2=a_df[['genus', 'date', 'Total_O2_consumption']]
perml_O2.to_csv(outdir+'total_O2_consumption_per_ml_per_genus.csv')

## Figure 3 csv files to create

A. cell diameter (volume) versus rRNA
B. cell diameter (volume) versus mRNA
C. PCA of Alpha
D. PCA of Bacterio
E. Genus Respiration v rRNA with size reflecting rhodopsin color by the color key above.
F. violin plots of Rhodopsin per um^3 of biomass split by Respiration levels 
    I need to figure out how to calculate biovolume per cell.

cols=['GTDB_genus', 'median_estimated_diameter']
genus=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Supplemental_table_genus_statistics_light_metabolism.csv',
                 usecols=cols)
genus=genus.rename(columns={'GTDB_genus':'genus'})
genus

cols=['genus', 'date', 'mRNA_transcripts_per_cell', '16S_copies_per_cell', 'Weighted_avg_O2_consumed_per_cell']
RNA=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/rRNA/rRNA_O2_transcripts_stats_all_genera_no2017.csv', 
                index_col=0, usecols=cols)

cols=['proteorhodopsin_transcripts_per_cell', 'genus', 'date']
rhod=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/rRNA/rRNA_O2_transcripts_stats_all_genera.csv', 
                 usecols=cols)


RNA=RNA.merge(genus, on='genus', how='left')
RNA=RNA.merge(rhod, on=['genus', 'date'], how='left')
RNA=RNA[RNA['genus'].isin(five_SAG)]
RNA

mRNA=RNA[['genus', 'date', 'mRNA_transcripts_per_cell', 'median_estimated_diameter']].copy()
mRNA=mRNA.dropna(how='any')
mRNA.to_csv(outdir+'Fig3_mRNA_diameter.csv')
mRNA

O2mRNA=RNA[['genus', 'date', 'mRNA_transcripts_per_cell', 'median_estimated_diameter', 'Weighted_avg_O2_consumed_per_cell']].copy()
O2mRNA=O2mRNA.dropna(how='any')
O2mRNA=O2mRNA[O2mRNA['Weighted_avg_O2_consumed_per_cell']>0.0025]
O2mRNA.to_csv(outdir+'Fig3_O2_genera_mRNA_diameter.csv')
O2mRNA

rRNA=RNA[['genus', 'date', '16S_copies_per_cell', 'median_estimated_diameter']].copy()
rRNA=rRNA.dropna(how='any')
rRNA.to_csv(outdir+'Fig3_rRNA_diameter.csv')
rRNA

O2rRNA=RNA[['genus', 'date', '16S_copies_per_cell', 'median_estimated_diameter', 'Weighted_avg_O2_consumed_per_cell']].copy()
O2rRNA=O2rRNA.dropna(how='any')
O2rRNA=O2rRNA[O2rRNA['Weighted_avg_O2_consumed_per_cell']>0.0025]
O2rRNA.to_csv(outdir+'Fig3_O2_genera_rRNA_diameter.csv')
O2rRNA

Rhod=RNA[['genus', 'date', '16S_copies_per_cell', 'Weighted_avg_O2_consumed_per_cell', 'proteorhodopsin_transcripts_per_cell']].copy()
Rhod=Rhod.dropna(how='any')
Rhod.to_csv(outdir+'Fig3E_rRNA_O2_rhodhopsin.csv')
Rhod

# FACS dot plots


cols=['name', 'Plate', 'Predicted_respiration_rate_fmol_O2perhr', 'GTDB_classification', 'gate', 'FACS_date', 'Sample_collection_Date', 'GTDBtk_classification']
SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                usecols=cols)

SAGs=SAGs[SAGs['FACS_date']==SAGs['Sample_collection_Date']]
SAGs=SAGs[SAGs['gate']=='Respiring cells (RSG-positive)']
SAGs[['domain', 'phyla', 'class', 'order', 'family', 'genus', 'species']]=SAGs.GTDBtk_classification.str.split(';', expand=True)

FACS_dot_plots=SAGs[['name', 'class', 'Plate']]

FACS_dot_plots['class']=FACS_dot_plots['class'].str[3:]

Classes=['Alphaproteobacteria', 'Gammaproteobacteria', 'Planctomycetes', 'Poseidoniia', 'Nitrososphaeria', 'Verrucomicrobiae', 'Acidimicrobiia',
         'Actinomycetia', 'Marinisomatia', 'Kapabacteria', 'Bacteroidia', 'Cyanobacteriia', 'UBA1135', 'Campylobacteria']

FACS_dot_plots['group']=np.where(FACS_dot_plots['class'].isin(Classes), FACS_dot_plots['class'], 'Other')
FACS_dot_plots

plates=FACS_dot_plots['Plate'].to_list()
plates=list(set(plates))
plates
for i in plates:
    df=FACS_dot_plots.loc[FACS_dot_plots['Plate'] == i ]
    df=df.drop(columns=['class', 'Plate'])
    i=str(i)
    df.to_csv(outdir+i+'_file_for_FACS_plot.csv', index=False)

df

# FACS dot plots


cols=['name', 'Plate', 'Predicted_respiration_rate_fmol_O2perhr', 'GTDB_classification', 'gate', 'FACS_date', 'Sample_collection_Date', 'GTDBtk_classification']
SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                usecols=cols)

SAGs=SAGs[SAGs['FACS_date']==SAGs['Sample_collection_Date']]
SAGs=SAGs[SAGs['gate']=='Respiring cells (RSG-positive)']
SAGs[['domain', 'phyla', 'class', 'order', 'family', 'genus', 'species']]=SAGs.GTDBtk_classification.str.split(';', expand=True)

FACS_dot_plots=SAGs[['name', 'genus', 'Plate', 'order']]

FACS_dot_plots['genus']=FACS_dot_plots['genus'].str[3:]
FACS_dot_plots['order']=FACS_dot_plots['order'].str[3:]

Classes=['ASP10-02a', "Flavobacteriales"]

FACS_dot_plots['group']=np.where(FACS_dot_plots['genus'].isin(Classes), FACS_dot_plots['genus'],
                                 np.where(FACS_dot_plots['order'].isin(Classes), FACS_dot_plots['order'],'Other'))



plates=FACS_dot_plots['Plate'].to_list()
plates=list(set(plates))
plates
for i in plates:
    df=FACS_dot_plots.loc[FACS_dot_plots['Plate'] == i ]
    df=df.drop(columns=['genus', 'Plate', 'order'])
    i=str(i)
    df.to_csv(outdir+i+'_ASP10_file_for_FACS_plot.csv', index=False)

df

cats=FACS_dot_plots['group'].to_list()
cats=list(set(cats))
cats


cols=['name', 'Plate', 'Predicted_respiration_rate_fmol_O2perhr', 'GTDB_classification', 'gate', 'FACS_date', 'Sample_collection_Date', 'GTDBtk_classification']
SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
                usecols=cols)

SAGs=SAGs[SAGs['FACS_date']==SAGs['Sample_collection_Date']]
SAGs=SAGs[SAGs['gate']=='All prokaryotes (SYTO-9-positive)']
SAGs[['domain', 'phyla', 'class', 'order', 'family', 'genus', 'species']]=SAGs.GTDBtk_classification.str.split(';', expand=True)

FACS_dot_plots=SAGs[['name', 'genus', 'Plate', 'order']]

FACS_dot_plots['genus']=FACS_dot_plots['genus'].str[3:]
FACS_dot_plots['order']=FACS_dot_plots['order'].str[3:]

Classes=['ASP10-02a', "Flavobacteriales"]

FACS_dot_plots['group']=np.where(FACS_dot_plots['genus'].isin(Classes), FACS_dot_plots['genus'],
                                 np.where(FACS_dot_plots['order'].isin(Classes), FACS_dot_plots['order'],'Other'))



plates=FACS_dot_plots['Plate'].to_list()
plates=list(set(plates))
plates
for i in plates:
    df=FACS_dot_plots.loc[FACS_dot_plots['Plate'] == i ]
    df=df.drop(columns=['genus', 'Plate', 'order'])
    i=str(i)
    df.to_csv(outdir+i+'_Syto9_SAGs_ASP10_file_for_FACS_plot.csv', index=False)

df

cats=FACS_dot_plots['group'].to_list()
cats=list(set(cats))
cats



vperc=pd.read_csv('/mnt/scgc/simon/microg2p/GoM_Viral_detection/Summary_files/All_days_virus_percentage.csv', index_col=0)
vperc['date']=vperc['date']+20000000

vO2=RNA[['genus', 'date', 'Weighted_avg_O2_consumed_per_cell']].copy()
vO2=vO2.drop_duplicates()
vperc=vperc.merge(vO2, on=['genus', 'date'], how='left')
vperc.to_csv(outdir+'All_days_avg_O2_consumption_and_abundance_virus_percentage.csv')
vperc
