==> 98_Misc.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import sys
import csv
from pathlib import Path
import seaborn as sns
sns.set()
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
from collections import defaultdict
import os.path as op
import glob
from csv import reader

os.chdir('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/')
outdir='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/'
images='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/'
matplotlib.__version__

rerun on 2021-12-07 after correcting for salinity of media samples

stats=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Genus_statistics.csv', index_col=0)
taxonomy=stats[['GTDB_genus', 'Phyla', 'Class', 'Order', 'Family', 'Genus']]
taxonomy=taxonomy.rename(columns={'GTDB_genus':'genus'})
family=taxonomy[['genus', 'Family']]
taxonomy


abundance=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/avg_O2_consumption_and_abundance_per_day_melt.csv', index_col=0)
abundance=abundance.sort_values(by='avg_fmol_O2_consumed', ascending=False)
abun_2018=abundance[abundance['date']==20181030]
abun_201904=abundance[abundance['date']==20190402]
abun_201907=abundance[abundance['date']==20190709]

abun_2018['percent']=(abun_2018['cells_per_ml']/abun_2018['cells_per_ml'].sum())*100
abun_2018=abun_2018.merge(taxonomy, how='left', on='genus' )

abun_201904['percent']=(abun_201904['cells_per_ml']/abun_201904['cells_per_ml'].sum())*100
abun_201904=abun_201904.merge(taxonomy, how='left', on='genus' )

abun_201907['percent']=(abun_201907['cells_per_ml']/abun_201907['cells_per_ml'].sum())*100
abun_201907=abun_201907.merge(taxonomy, how='left', on='genus' )

abundance

rhod18=abun_2018[abun_2018['Family']=='Rhodobacteraceae']

rhod18.sum()

rhod1904=abun_201904[abun_201904['Family']=='Rhodobacteraceae']
rhod1904.sum()

rhod1907=abun_201907[abun_201907['Family']=='Rhodobacteraceae']
rhod1907.sum()

flav=abun_2018[abun_2018['Family']=='Flavobacteriaceae']
flav=flav.append(abun_201904[abun_201904['Family']=='Flavobacteriaceae'])
flav=flav.append(abun_201907[abun_201907['Family']=='Flavobacteriaceae'])
flav

Total=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption/O2_consumed_by_cells_above_and_below_detection_limit_per_genus.csv',
                 index_col=0)
Total=Total[Total['date']!= 20171102]
Total=Total[Total['cells/ml']!=0]
Total=Total.drop('fraction', axis=1)
Total=Total.groupby(['genus', 'date']).sum()
Total.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_analysis/O2_consumption/Total_O2_consumption_and_abundance_per_genus.csv')
Total=Total.reset_index()
Total=Total.merge(family, on='genus', how='left')
Total=Total.rename(columns={'fmol_O2_consumed_per_ml_per_hr':'total_fmol_O2_consumed_per_ml_per_hr'})
Total

Total_2018=Total[Total['date']==20181030]
Total_2018=Total_2018.sort_values(by='total_fmol_O2_consumed_per_ml_per_hr', ascending=False)
Last=Total_2018[Total_2018['genus'] == 'Unclassified']
Total_2018=Total_2018[Total_2018['genus']!= 'Unclassified']
Total_2018=Total_2018.append(Last)
Total_2018['O2_perc']=(Total_2018['total_fmol_O2_consumed_per_ml_per_hr']/Total_2018['total_fmol_O2_consumed_per_ml_per_hr'].sum())*100
Total_2018['cells_perc']=(Total_2018['cells/ml']/Total_2018['cells/ml'].sum())*100

Rhod_2018=Total_2018[Total_2018['Family']== 'Rhodobacteraceae']
print(Rhod_2018[['O2_perc', 'cells_perc']].sum())

Total_2018['cumultative_cellperc']=Total_2018['cells_perc'].cumsum()
Total_2018['cumultative_O2_consumed']=Total_2018['O2_perc'].cumsum()

Total_2018


Rhod_2018

Total_201904=Total[Total['date']==20190402]
Total_201904=Total_201904.sort_values(by='total_fmol_O2_consumed_per_ml_per_hr', ascending=False)
Last=Total_201904[Total_201904['genus'] == 'Unclassified']
Total_201904=Total_201904[Total_201904['genus']!= 'Unclassified']
Total_201904=Total_201904.append(Last)
Total_201904['O2_perc']=(Total_201904['total_fmol_O2_consumed_per_ml_per_hr']/Total_201904['total_fmol_O2_consumed_per_ml_per_hr'].sum())*100
Total_201904['cells_perc']=(Total_201904['cells/ml']/Total_201904['cells/ml'].sum())*100

Rhod_201904=Total_201904[Total_201904['Family']== 'Rhodobacteraceae']
print("Rhodobacteraceae \n", Rhod_201904[['O2_perc', 'cells_perc']].sum())

Flav_201904=Total_201904[Total_201904['Family']== 'Flavobacteriaceae']
print("Flavobacteriaceae \n", Flav_201904[['O2_perc', 'cells_perc']].sum())

Total_201904['cumultative_cellperc']=Total_201904['cells_perc'].cumsum()
Total_201904['cumultative_O2_consumed']=Total_201904['O2_perc'].cumsum()


Total_201904


Total_201907=Total[Total['date']==20190709]
Total_201907=Total_201907.sort_values(by='total_fmol_O2_consumed_per_ml_per_hr', ascending=False)
Last=Total_201907[Total_201907['genus'] == 'Unclassified']
Total_201907=Total_201907[Total_201907['genus']!= 'Unclassified']
Total_201907=Total_201907.append(Last)
Total_201907['O2_perc']=(Total_201907['total_fmol_O2_consumed_per_ml_per_hr']/Total_201907['total_fmol_O2_consumed_per_ml_per_hr'].sum())*100
Total_201907['cells_perc']=(Total_201907['cells/ml']/Total_201907['cells/ml'].sum())*100

Rhod_201907=Total_201907[Total_201907['Family']== 'Rhodobacteraceae']
print(Rhod_201907[['O2_perc', 'cells_perc']].sum())

Total_201907['cumultative_cellperc']=Total_201907['cells_perc'].cumsum()
Total_201907['cumultative_O2_consumed']=Total_201907['O2_perc'].cumsum()

Total_201907


Rhod_201907

SAGs=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv')
print(SAGs['GTDB_classification'].nunique())
SAGs

RSG=SAGs[SAGs['gate']=='Respiring cells (RSG-positive)']
RSG=RSG[RSG['Sample_collection_Date'] == RSG['FACS_date']]
print(RSG['GTDB_classification'].nunique())



print(SAGs['gate'].unique())

with open("/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/Genera_with_5+SAGs.txt") as file:
    fiveGENs_list = [line.strip() for line in file]
fiveGENs_list.append('Unclassified')
SAG=pd.read_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv',
               index_col=0)

fiveSAG=SAG[SAG['GTDB_classification'].isin(fiveGENs_list)]
fiveSAG.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_5SAG_genera_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass.csv')
fiveSAG=fiveSAG[fiveSAG['gate']=='Respiring cells (RSG-positive)']
fiveSAG=fiveSAG[fiveSAG['Sample_collection_Date']==fiveSAG['FACS_date']]
dates=[181030, 190402, 190709]
fiveSAG=fiveSAG[fiveSAG['Sample_collection_Date'].isin(dates)]
fiveSAG.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_5SAG_genera_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass_Fresh_RSG_2018_2019_only.csv')


fiveSAG['Sample_collection_Date'].unique()


tree_SAG=SAG[SAG['gate']=='Respiring cells (RSG-positive)']
tree_SAG=tree_SAG[tree_SAG['Sample_collection_Date']==tree_SAG['FACS_date']]
dates=[181030, 190402, 190709]
tree_SAG=tree_SAG[tree_SAG['Sample_collection_Date'].isin(dates)]
tree_SAG.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_genera_1cell_20kb_decon_531normalized_predresp_rate_GTDBclass_Fresh_RSG_2018_2019_only.csv')
