==> 98a_Genome_copier_for_submission.ipynb <==
import pandas as pd
from pandas import DataFrame
import os
import os.path as op
import sys
import csv
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
plt.show()
import glob
import seaborn as sns
import shutil
import networkx as nx

# import my custom functions
sys.path.insert(0, '/home/jmunson-mcgee/')
from JMM_functions import *

pd.set_option("display.max_rows", 6)

outdir='/mnt/scgc/simon/microg2p/Data/GoM_uploads/SAG_contigs'
safe_make_dir(outdir)
os.chdir(outdir)
    
matplotlib.__version__

SAG_file='/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/All_GoM_SAGs_metadata.csv'


cols=['name', 'Predicted_respiration_rate_fmol_O2perhr', 'GTDB_classification', 'estimated_diameter', 'gate', 'Plate']
GoM_SAGs=pd.read_csv(SAG_file, index_col=0)

GoM_SAGs=GoM_SAGs[GoM_SAGs['final_assembly_length']>=20000]

# Get rid of the Cryopreserved SAGs that were not used in the read recruitment dataset
GoM_SAGs=GoM_SAGs[GoM_SAGs['Plate']!='AM-379']

SAGs=GoM_SAGs['name'].tolist()
plates=GoM_SAGs['Plate'].unique().tolist()
print(len(SAGs))
GoM_SAGs

ls1=['AM-377-I17', 'AM-384-D14', 'AM-382-D14']
test=GoM_SAGs[GoM_SAGs['name'].isin(ls1)]
test

contig_files=glob.glob('/mnt/scgc/simon/microg2p/Data/SAG_contigs/cryo_RSG_contigs/*.fasta')+\
glob.glob('/mnt/scgc/simon/microg2p/Data/SAG_contigs/2021-08_SAG_contigs/*.fasta')+\
glob.glob('/mnt/scgc/simon/microg2p/Data/SAG_contigs/GoM/1*/*.fasta')+\
glob.glob('/mnt/scgc/simon/microg2p/Data/SAG_contigs/Poseidon/*.fasta')+\
glob.glob('/mnt/scgc/simon/microg2p/Data/SAG_contigs/Radprof/*.fasta')+\
glob.glob('/mnt/scgc/simon/microg2p/Data/SAG_contigs/AT42-11/*.fasta')+\
glob.glob('/mnt/scgc/simon/microg2p/Data/Herndl_WGS/contigs/AH-716*.fasta')
print(len(contig_files))

cp_files=[]

for i in SAGs:
    for file in contig_files:
        if i in file:
            cp_files.append(file)
print(len(cp_files))   


cp_files=list(set(cp_files))
len(cp_files)



cp_SAGs=[]
for i in cp_files:
    SAG=os.path.basename(i)
    SAG=SAG[:10]
    
    cp_SAGs.append(SAG)
print(len(cp_SAGs))

missing_SAGs=[]
for i in SAGs:
    if i not in cp_SAGs:
        missing_SAGs.append(i)
print(len(set(missing_SAGs)))


plates=[]
for i in missing_SAGs:
    p=i[:6]
    plates.append(p)
plates=list(set(plates))
plates


missing=GoM_SAGs[GoM_SAGs['name'].isin(missing_SAGs)]

missing
missing_SAGs=missing['name'].tolist()

missing


FINAL_SAGs=GoM_SAGs[~GoM_SAGs['name'].isin(missing_SAGs)]


FINAL_SAGs.to_csv('/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/Summary_files/FINAL_GoM_Metadata.csv')
# create a list of the final SAGs
SAGs=FINAL_SAGs['name'].tolist()
SAG_list='/mnt/scgc/simon/microg2p/Data/GoM_uploads/SAGlist.txt'
write_list(SAGs, SAG_list)

FINAL_SAGs


dest_contigs=glob.glob('/mnt/scgc/simon/microg2p/Data/GoM_uploads/SAG_contigs/*.fasta')
dest_files=[]
for i in dest_contigs:
    name=os.path.basename(i)
    dest_files.append(name)
    
print(len(dest_files))

print(len(cp_files))
for file in cp_files:
    name=os.path.basename(file)
    if name not in dest_files:        
        shutil.copy(file, outdir)
#        print(name)
#         print(file)
!ls |wc -l

! wc -l /mnt/scgc/simon/microg2p/Data/GoM_uploads/SAGlist.txt

os.chdir('/mnt/scgc/simon/microg2p/Data/GoM_uploads/SAG_contigs/')
SAG_files=glob.glob('*.fasta')
filelist='/mnt/scgc/simon/microg2p/Data/GoM_uploads/SAG_filename.txt'
write_list(SAG_files, filelist)
!wc -l /mnt/scgc/simon/microg2p/Data/GoM_uploads/SAG_filename.txt
