==> 99_Fig3_plots_R.ipynb <==
library(randomcoloR)
library(tidyverse)
library(scales)

`%notin%` <- Negate(`%in%`)

setwd("/mnt/scgc/simon/microg2p/analyses/20210325_GoM_recluster/20210325_GoM_recluster_images/csv_files_for_figures")


#read input data
mRNA <- read.csv(file="Fig3_mRNA_diameter.csv", header=TRUE, stringsAsFactors = FALSE)

rRNA <- read.csv(file="Fig3_rRNA_diameter.csv", header=TRUE, stringsAsFactors = FALSE)

rhod <- read.csv(file="Fig3F_rhodopsin_per_µm.csv", header=TRUE, stringsAsFactors = FALSE)


class <- c("Alphaproteobacteria", "Gammaproteobacteria", "Planctomycetes", "Poseidoniia", "Nitrososphaeria", "Verrucomicrobiae", 
           "Acidimicrobiia", "Actinomycetia", "Actinobacteria", "Marinisomatia", "Kapabacteria" ,"Bacteroidia", "Cyanobacteriia", 
           "UBA1135", "Campylobacteria")
class_colors <- c('#2f4f4f', '#8b4513', '#6b8e23', '#000080', '#ff0000', '#00ced1',
                 '#ffa500', '#ffff00', '#c71585', '#00fa9a', '#0000ff', '#00ff00', '#ff00ff', 
                  '#6495ed', '#ffe4b5')
